more work on isolating antlr imports
