+++ .gitignore
@@ -26,6 +26,9 @@
 # Generated dynamic libraries in the polys directory
 /sympy/polys/*.so
 
+# Generated atlr assets
+/sympy/parsing/_latex/_antlr_raw
+
 # File generated by setup.py using MANIFEST.in
 MANIFEST
