+++ MANIFEST.in
@@ -5,7 +5,6 @@ recursive-include examples *.py README
 
 include sympy/utilities/mathml/data/*.xsl
 include sympy/parsing/_latex/*.g4
-include sympy/parsing/_latex/*.tokens
 include sympy/parsing/_latex/LICENSE.txt
 
 include LICENSE
