+++ sympy/parsing/_latex/_antlr/latexlexer.py
@@ -1,0 +1,381 @@
+def _importer():
+    try:
+        # Generated from LaTeX.g4 by ANTLR 4.7.1
+        # encoding: utf-8
+        from antlr4 import (
+            Token,
+            InputStream,
+            FileStream,
+            error,
+            tree,
+            Utils,
+            RuleContext,
+            ParserRuleContext,
+            Recognizer,
+            BufferedTokenStream,
+            TokenStream,
+            CommonTokenFactory,
+            atn,
+            IntervalSet,
+            PredictionContext,
+            dfa,
+            Lexer,
+            CommonTokenStream,
+            ListTokenSource,
+            Parser,
+            DFA,
+            ATN,
+            ATNDeserializer,
+            LexerATNSimulator,
+            ParserATNSimulator,
+            PredictionMode,
+            PredictionContextCache,
+            ParseTreeListener,
+            ParseTreeVisitor,
+            ParseTreeWalker,
+            TerminalNode,
+            ErrorNode,
+            RuleNode,
+            RecognitionException,
+            IllegalStateException,
+            NoViableAltException,
+            BailErrorStrategy,
+            DiagnosticErrorListener,
+            str_list
+        )
+        from io import StringIO
+        import sys
+        
+        
+        def serializedATN():
+            with StringIO() as buf:
+                buf.write(u"\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\2")
+                buf.write(u";\u01e8\b\1\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4")
+                buf.write(u"\7\t\7\4\b\t\b\4\t\t\t\4\n\t\n\4\13\t\13\4\f\t\f\4\r")
+                buf.write(u"\t\r\4\16\t\16\4\17\t\17\4\20\t\20\4\21\t\21\4\22\t\22")
+                buf.write(u"\4\23\t\23\4\24\t\24\4\25\t\25\4\26\t\26\4\27\t\27\4")
+                buf.write(u"\30\t\30\4\31\t\31\4\32\t\32\4\33\t\33\4\34\t\34\4\35")
+                buf.write(u"\t\35\4\36\t\36\4\37\t\37\4 \t \4!\t!\4\"\t\"\4#\t#\4")
+                buf.write(u"$\t$\4%\t%\4&\t&\4\'\t\'\4(\t(\4)\t)\4*\t*\4+\t+\4,\t")
+                buf.write(u",\4-\t-\4.\t.\4/\t/\4\60\t\60\4\61\t\61\4\62\t\62\4\63")
+                buf.write(u"\t\63\4\64\t\64\4\65\t\65\4\66\t\66\4\67\t\67\48\t8\4")
+                buf.write(u"9\t9\4:\t:\4;\t;\4<\t<\3\2\3\2\3\3\6\3}\n\3\r\3\16\3")
+                buf.write(u"~\3\3\3\3\3\4\3\4\3\5\3\5\3\6\3\6\3\7\3\7\3\b\3\b\3\t")
+                buf.write(u"\3\t\3\n\3\n\3\13\3\13\3\f\3\f\3\r\3\r\3\16\3\16\3\17")
+                buf.write(u"\3\17\3\17\3\17\3\17\3\20\3\20\3\20\3\20\3\20\3\20\3")
+                buf.write(u"\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20")
+                buf.write(u"\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3")
+                buf.write(u"\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20")
+                buf.write(u"\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3\20\3")
+                buf.write(u"\20\3\20\3\20\3\20\3\20\3\20\3\20\5\20\u00d5\n\20\3\21")
+                buf.write(u"\3\21\3\21\3\21\3\21\3\22\3\22\3\22\3\22\3\22\3\23\3")
+                buf.write(u"\23\3\23\3\23\3\23\3\23\3\24\3\24\3\24\3\24\3\24\3\25")
+                buf.write(u"\3\25\3\25\3\25\3\26\3\26\3\26\3\26\3\26\3\27\3\27\3")
+                buf.write(u"\27\3\27\3\27\3\30\3\30\3\30\3\30\3\30\3\31\3\31\3\31")
+                buf.write(u"\3\31\3\31\3\32\3\32\3\32\3\32\3\32\3\33\3\33\3\33\3")
+                buf.write(u"\33\3\33\3\34\3\34\3\34\3\34\3\34\3\34\3\34\3\34\3\35")
+                buf.write(u"\3\35\3\35\3\35\3\35\3\35\3\35\3\35\3\36\3\36\3\36\3")
+                buf.write(u"\36\3\36\3\36\3\36\3\36\3\37\3\37\3\37\3\37\3\37\3\37")
+                buf.write(u"\3\37\3\37\3 \3 \3 \3 \3 \3 \3 \3 \3!\3!\3!\3!\3!\3!")
+                buf.write(u"\3!\3!\3\"\3\"\3\"\3\"\3\"\3\"\3#\3#\3#\3#\3#\3#\3$\3")
+                buf.write(u"$\3$\3$\3$\3$\3%\3%\3%\3%\3%\3%\3%\3%\3&\3&\3&\3&\3&")
+                buf.write(u"\3&\3&\3&\3\'\3\'\3\'\3\'\3\'\3\'\3\'\3\'\3(\3(\3(\3")
+                buf.write(u"(\3(\3(\3)\3)\3)\3)\3)\3)\3)\3*\3*\3*\3*\3*\3*\3+\3+")
+                buf.write(u"\3+\3+\3+\3,\3,\3,\3,\3,\3,\3-\3-\3-\3-\3-\3-\3-\3-\3")
+                buf.write(u".\3.\3/\3/\3\60\3\60\3\61\3\61\3\62\3\62\7\62\u0198\n")
+                buf.write(u"\62\f\62\16\62\u019b\13\62\3\62\3\62\3\62\6\62\u01a0")
+                buf.write(u"\n\62\r\62\16\62\u01a1\5\62\u01a4\n\62\3\63\3\63\3\64")
+                buf.write(u"\3\64\3\65\6\65\u01ab\n\65\r\65\16\65\u01ac\3\65\3\65")
+                buf.write(u"\3\65\3\65\3\65\7\65\u01b4\n\65\f\65\16\65\u01b7\13\65")
+                buf.write(u"\3\65\7\65\u01ba\n\65\f\65\16\65\u01bd\13\65\3\65\3\65")
+                buf.write(u"\3\65\3\65\3\65\7\65\u01c4\n\65\f\65\16\65\u01c7\13\65")
+                buf.write(u"\3\65\3\65\6\65\u01cb\n\65\r\65\16\65\u01cc\5\65\u01cf")
+                buf.write(u"\n\65\3\66\3\66\3\67\3\67\38\38\38\38\38\39\39\3:\3:")
+                buf.write(u"\3:\3:\3:\3;\3;\3<\3<\6<\u01e5\n<\r<\16<\u01e6\3\u0199")
+                buf.write(u"\2=\3\3\5\4\7\5\t\6\13\7\r\b\17\t\21\n\23\13\25\f\27")
+                buf.write(u"\r\31\16\33\17\35\20\37\21!\22#\23%\24\'\25)\26+\27-")
+                buf.write(u"\30/\31\61\32\63\33\65\34\67\359\36;\37= ?!A\"C#E$G%")
+                buf.write(u"I&K\'M(O)Q*S+U,W-Y.[/]\60_\61a\2c\62e\63g\2i\64k\65m")
+                buf.write(u"\66o\67q8s9u:w;\3\2\5\5\2\13\f\17\17\"\"\4\2C\\c|\3\2")
+                buf.write(u"\62;\2\u01f4\2\3\3\2\2\2\2\5\3\2\2\2\2\7\3\2\2\2\2\t")
+                buf.write(u"\3\2\2\2\2\13\3\2\2\2\2\r\3\2\2\2\2\17\3\2\2\2\2\21\3")
+                buf.write(u"\2\2\2\2\23\3\2\2\2\2\25\3\2\2\2\2\27\3\2\2\2\2\31\3")
+                buf.write(u"\2\2\2\2\33\3\2\2\2\2\35\3\2\2\2\2\37\3\2\2\2\2!\3\2")
+                buf.write(u"\2\2\2#\3\2\2\2\2%\3\2\2\2\2\'\3\2\2\2\2)\3\2\2\2\2+")
+                buf.write(u"\3\2\2\2\2-\3\2\2\2\2/\3\2\2\2\2\61\3\2\2\2\2\63\3\2")
+                buf.write(u"\2\2\2\65\3\2\2\2\2\67\3\2\2\2\29\3\2\2\2\2;\3\2\2\2")
+                buf.write(u"\2=\3\2\2\2\2?\3\2\2\2\2A\3\2\2\2\2C\3\2\2\2\2E\3\2\2")
+                buf.write(u"\2\2G\3\2\2\2\2I\3\2\2\2\2K\3\2\2\2\2M\3\2\2\2\2O\3\2")
+                buf.write(u"\2\2\2Q\3\2\2\2\2S\3\2\2\2\2U\3\2\2\2\2W\3\2\2\2\2Y\3")
+                buf.write(u"\2\2\2\2[\3\2\2\2\2]\3\2\2\2\2_\3\2\2\2\2c\3\2\2\2\2")
+                buf.write(u"e\3\2\2\2\2i\3\2\2\2\2k\3\2\2\2\2m\3\2\2\2\2o\3\2\2\2")
+                buf.write(u"\2q\3\2\2\2\2s\3\2\2\2\2u\3\2\2\2\2w\3\2\2\2\3y\3\2\2")
+                buf.write(u"\2\5|\3\2\2\2\7\u0082\3\2\2\2\t\u0084\3\2\2\2\13\u0086")
+                buf.write(u"\3\2\2\2\r\u0088\3\2\2\2\17\u008a\3\2\2\2\21\u008c\3")
+                buf.write(u"\2\2\2\23\u008e\3\2\2\2\25\u0090\3\2\2\2\27\u0092\3\2")
+                buf.write(u"\2\2\31\u0094\3\2\2\2\33\u0096\3\2\2\2\35\u0098\3\2\2")
+                buf.write(u"\2\37\u00d4\3\2\2\2!\u00d6\3\2\2\2#\u00db\3\2\2\2%\u00e0")
+                buf.write(u"\3\2\2\2\'\u00e6\3\2\2\2)\u00eb\3\2\2\2+\u00ef\3\2\2")
+                buf.write(u"\2-\u00f4\3\2\2\2/\u00f9\3\2\2\2\61\u00fe\3\2\2\2\63")
+                buf.write(u"\u0103\3\2\2\2\65\u0108\3\2\2\2\67\u010d\3\2\2\29\u0115")
+                buf.write(u"\3\2\2\2;\u011d\3\2\2\2=\u0125\3\2\2\2?\u012d\3\2\2\2")
+                buf.write(u"A\u0135\3\2\2\2C\u013d\3\2\2\2E\u0143\3\2\2\2G\u0149")
+                buf.write(u"\3\2\2\2I\u014f\3\2\2\2K\u0157\3\2\2\2M\u015f\3\2\2\2")
+                buf.write(u"O\u0167\3\2\2\2Q\u016d\3\2\2\2S\u0174\3\2\2\2U\u017a")
+                buf.write(u"\3\2\2\2W\u017f\3\2\2\2Y\u0185\3\2\2\2[\u018d\3\2\2\2")
+                buf.write(u"]\u018f\3\2\2\2_\u0191\3\2\2\2a\u0193\3\2\2\2c\u0195")
+                buf.write(u"\3\2\2\2e\u01a5\3\2\2\2g\u01a7\3\2\2\2i\u01ce\3\2\2\2")
+                buf.write(u"k\u01d0\3\2\2\2m\u01d2\3\2\2\2o\u01d4\3\2\2\2q\u01d9")
+                buf.write(u"\3\2\2\2s\u01db\3\2\2\2u\u01e0\3\2\2\2w\u01e2\3\2\2\2")
+                buf.write(u"yz\7.\2\2z\4\3\2\2\2{}\t\2\2\2|{\3\2\2\2}~\3\2\2\2~|")
+                buf.write(u"\3\2\2\2~\177\3\2\2\2\177\u0080\3\2\2\2\u0080\u0081\b")
+                buf.write(u"\3\2\2\u0081\6\3\2\2\2\u0082\u0083\7-\2\2\u0083\b\3\2")
+                buf.write(u"\2\2\u0084\u0085\7/\2\2\u0085\n\3\2\2\2\u0086\u0087\7")
+                buf.write(u",\2\2\u0087\f\3\2\2\2\u0088\u0089\7\61\2\2\u0089\16\3")
+                buf.write(u"\2\2\2\u008a\u008b\7*\2\2\u008b\20\3\2\2\2\u008c\u008d")
+                buf.write(u"\7+\2\2\u008d\22\3\2\2\2\u008e\u008f\7}\2\2\u008f\24")
+                buf.write(u"\3\2\2\2\u0090\u0091\7\177\2\2\u0091\26\3\2\2\2\u0092")
+                buf.write(u"\u0093\7]\2\2\u0093\30\3\2\2\2\u0094\u0095\7_\2\2\u0095")
+                buf.write(u"\32\3\2\2\2\u0096\u0097\7~\2\2\u0097\34\3\2\2\2\u0098")
+                buf.write(u"\u0099\7^\2\2\u0099\u009a\7n\2\2\u009a\u009b\7k\2\2\u009b")
+                buf.write(u"\u009c\7o\2\2\u009c\36\3\2\2\2\u009d\u009e\7^\2\2\u009e")
+                buf.write(u"\u009f\7v\2\2\u009f\u00d5\7q\2\2\u00a0\u00a1\7^\2\2\u00a1")
+                buf.write(u"\u00a2\7t\2\2\u00a2\u00a3\7k\2\2\u00a3\u00a4\7i\2\2\u00a4")
+                buf.write(u"\u00a5\7j\2\2\u00a5\u00a6\7v\2\2\u00a6\u00a7\7c\2\2\u00a7")
+                buf.write(u"\u00a8\7t\2\2\u00a8\u00a9\7t\2\2\u00a9\u00aa\7q\2\2\u00aa")
+                buf.write(u"\u00d5\7y\2\2\u00ab\u00ac\7^\2\2\u00ac\u00ad\7T\2\2\u00ad")
+                buf.write(u"\u00ae\7k\2\2\u00ae\u00af\7i\2\2\u00af\u00b0\7j\2\2\u00b0")
+                buf.write(u"\u00b1\7v\2\2\u00b1\u00b2\7c\2\2\u00b2\u00b3\7t\2\2\u00b3")
+                buf.write(u"\u00b4\7t\2\2\u00b4\u00b5\7q\2\2\u00b5\u00d5\7y\2\2\u00b6")
+                buf.write(u"\u00b7\7^\2\2\u00b7\u00b8\7n\2\2\u00b8\u00b9\7q\2\2\u00b9")
+                buf.write(u"\u00ba\7p\2\2\u00ba\u00bb\7i\2\2\u00bb\u00bc\7t\2\2\u00bc")
+                buf.write(u"\u00bd\7k\2\2\u00bd\u00be\7i\2\2\u00be\u00bf\7j\2\2\u00bf")
+                buf.write(u"\u00c0\7v\2\2\u00c0\u00c1\7c\2\2\u00c1\u00c2\7t\2\2\u00c2")
+                buf.write(u"\u00c3\7t\2\2\u00c3\u00c4\7q\2\2\u00c4\u00d5\7y\2\2\u00c5")
+                buf.write(u"\u00c6\7^\2\2\u00c6\u00c7\7N\2\2\u00c7\u00c8\7q\2\2\u00c8")
+                buf.write(u"\u00c9\7p\2\2\u00c9\u00ca\7i\2\2\u00ca\u00cb\7t\2\2\u00cb")
+                buf.write(u"\u00cc\7k\2\2\u00cc\u00cd\7i\2\2\u00cd\u00ce\7j\2\2\u00ce")
+                buf.write(u"\u00cf\7v\2\2\u00cf\u00d0\7c\2\2\u00d0\u00d1\7t\2\2\u00d1")
+                buf.write(u"\u00d2\7t\2\2\u00d2\u00d3\7q\2\2\u00d3\u00d5\7y\2\2\u00d4")
+                buf.write(u"\u009d\3\2\2\2\u00d4\u00a0\3\2\2\2\u00d4\u00ab\3\2\2")
+                buf.write(u"\2\u00d4\u00b6\3\2\2\2\u00d4\u00c5\3\2\2\2\u00d5 \3\2")
+                buf.write(u"\2\2\u00d6\u00d7\7^\2\2\u00d7\u00d8\7k\2\2\u00d8\u00d9")
+                buf.write(u"\7p\2\2\u00d9\u00da\7v\2\2\u00da\"\3\2\2\2\u00db\u00dc")
+                buf.write(u"\7^\2\2\u00dc\u00dd\7u\2\2\u00dd\u00de\7w\2\2\u00de\u00df")
+                buf.write(u"\7o\2\2\u00df$\3\2\2\2\u00e0\u00e1\7^\2\2\u00e1\u00e2")
+                buf.write(u"\7r\2\2\u00e2\u00e3\7t\2\2\u00e3\u00e4\7q\2\2\u00e4\u00e5")
+                buf.write(u"\7f\2\2\u00e5&\3\2\2\2\u00e6\u00e7\7^\2\2\u00e7\u00e8")
+                buf.write(u"\7n\2\2\u00e8\u00e9\7q\2\2\u00e9\u00ea\7i\2\2\u00ea(")
+                buf.write(u"\3\2\2\2\u00eb\u00ec\7^\2\2\u00ec\u00ed\7n\2\2\u00ed")
+                buf.write(u"\u00ee\7p\2\2\u00ee*\3\2\2\2\u00ef\u00f0\7^\2\2\u00f0")
+                buf.write(u"\u00f1\7u\2\2\u00f1\u00f2\7k\2\2\u00f2\u00f3\7p\2\2\u00f3")
+                buf.write(u",\3\2\2\2\u00f4\u00f5\7^\2\2\u00f5\u00f6\7e\2\2\u00f6")
+                buf.write(u"\u00f7\7q\2\2\u00f7\u00f8\7u\2\2\u00f8.\3\2\2\2\u00f9")
+                buf.write(u"\u00fa\7^\2\2\u00fa\u00fb\7v\2\2\u00fb\u00fc\7c\2\2\u00fc")
+                buf.write(u"\u00fd\7p\2\2\u00fd\60\3\2\2\2\u00fe\u00ff\7^\2\2\u00ff")
+                buf.write(u"\u0100\7e\2\2\u0100\u0101\7u\2\2\u0101\u0102\7e\2\2\u0102")
+                buf.write(u"\62\3\2\2\2\u0103\u0104\7^\2\2\u0104\u0105\7u\2\2\u0105")
+                buf.write(u"\u0106\7g\2\2\u0106\u0107\7e\2\2\u0107\64\3\2\2\2\u0108")
+                buf.write(u"\u0109\7^\2\2\u0109\u010a\7e\2\2\u010a\u010b\7q\2\2\u010b")
+                buf.write(u"\u010c\7v\2\2\u010c\66\3\2\2\2\u010d\u010e\7^\2\2\u010e")
+                buf.write(u"\u010f\7c\2\2\u010f\u0110\7t\2\2\u0110\u0111\7e\2\2\u0111")
+                buf.write(u"\u0112\7u\2\2\u0112\u0113\7k\2\2\u0113\u0114\7p\2\2\u0114")
+                buf.write(u"8\3\2\2\2\u0115\u0116\7^\2\2\u0116\u0117\7c\2\2\u0117")
+                buf.write(u"\u0118\7t\2\2\u0118\u0119\7e\2\2\u0119\u011a\7e\2\2\u011a")
+                buf.write(u"\u011b\7q\2\2\u011b\u011c\7u\2\2\u011c:\3\2\2\2\u011d")
+                buf.write(u"\u011e\7^\2\2\u011e\u011f\7c\2\2\u011f\u0120\7t\2\2\u0120")
+                buf.write(u"\u0121\7e\2\2\u0121\u0122\7v\2\2\u0122\u0123\7c\2\2\u0123")
+                buf.write(u"\u0124\7p\2\2\u0124<\3\2\2\2\u0125\u0126\7^\2\2\u0126")
+                buf.write(u"\u0127\7c\2\2\u0127\u0128\7t\2\2\u0128\u0129\7e\2\2\u0129")
+                buf.write(u"\u012a\7e\2\2\u012a\u012b\7u\2\2\u012b\u012c\7e\2\2\u012c")
+                buf.write(u">\3\2\2\2\u012d\u012e\7^\2\2\u012e\u012f\7c\2\2\u012f")
+                buf.write(u"\u0130\7t\2\2\u0130\u0131\7e\2\2\u0131\u0132\7u\2\2\u0132")
+                buf.write(u"\u0133\7g\2\2\u0133\u0134\7e\2\2\u0134@\3\2\2\2\u0135")
+                buf.write(u"\u0136\7^\2\2\u0136\u0137\7c\2\2\u0137\u0138\7t\2\2\u0138")
+                buf.write(u"\u0139\7e\2\2\u0139\u013a\7e\2\2\u013a\u013b\7q\2\2\u013b")
+                buf.write(u"\u013c\7v\2\2\u013cB\3\2\2\2\u013d\u013e\7^\2\2\u013e")
+                buf.write(u"\u013f\7u\2\2\u013f\u0140\7k\2\2\u0140\u0141\7p\2\2\u0141")
+                buf.write(u"\u0142\7j\2\2\u0142D\3\2\2\2\u0143\u0144\7^\2\2\u0144")
+                buf.write(u"\u0145\7e\2\2\u0145\u0146\7q\2\2\u0146\u0147\7u\2\2\u0147")
+                buf.write(u"\u0148\7j\2\2\u0148F\3\2\2\2\u0149\u014a\7^\2\2\u014a")
+                buf.write(u"\u014b\7v\2\2\u014b\u014c\7c\2\2\u014c\u014d\7p\2\2\u014d")
+                buf.write(u"\u014e\7j\2\2\u014eH\3\2\2\2\u014f\u0150\7^\2\2\u0150")
+                buf.write(u"\u0151\7c\2\2\u0151\u0152\7t\2\2\u0152\u0153\7u\2\2\u0153")
+                buf.write(u"\u0154\7k\2\2\u0154\u0155\7p\2\2\u0155\u0156\7j\2\2\u0156")
+                buf.write(u"J\3\2\2\2\u0157\u0158\7^\2\2\u0158\u0159\7c\2\2\u0159")
+                buf.write(u"\u015a\7t\2\2\u015a\u015b\7e\2\2\u015b\u015c\7q\2\2\u015c")
+                buf.write(u"\u015d\7u\2\2\u015d\u015e\7j\2\2\u015eL\3\2\2\2\u015f")
+                buf.write(u"\u0160\7^\2\2\u0160\u0161\7c\2\2\u0161\u0162\7t\2\2\u0162")
+                buf.write(u"\u0163\7v\2\2\u0163\u0164\7c\2\2\u0164\u0165\7p\2\2\u0165")
+                buf.write(u"\u0166\7j\2\2\u0166N\3\2\2\2\u0167\u0168\7^\2\2\u0168")
+                buf.write(u"\u0169\7u\2\2\u0169\u016a\7s\2\2\u016a\u016b\7t\2\2\u016b")
+                buf.write(u"\u016c\7v\2\2\u016cP\3\2\2\2\u016d\u016e\7^\2\2\u016e")
+                buf.write(u"\u016f\7v\2\2\u016f\u0170\7k\2\2\u0170\u0171\7o\2\2\u0171")
+                buf.write(u"\u0172\7g\2\2\u0172\u0173\7u\2\2\u0173R\3\2\2\2\u0174")
+                buf.write(u"\u0175\7^\2\2\u0175\u0176\7e\2\2\u0176\u0177\7f\2\2\u0177")
+                buf.write(u"\u0178\7q\2\2\u0178\u0179\7v\2\2\u0179T\3\2\2\2\u017a")
+                buf.write(u"\u017b\7^\2\2\u017b\u017c\7f\2\2\u017c\u017d\7k\2\2\u017d")
+                buf.write(u"\u017e\7x\2\2\u017eV\3\2\2\2\u017f\u0180\7^\2\2\u0180")
+                buf.write(u"\u0181\7h\2\2\u0181\u0182\7t\2\2\u0182\u0183\7c\2\2\u0183")
+                buf.write(u"\u0184\7e\2\2\u0184X\3\2\2\2\u0185\u0186\7^\2\2\u0186")
+                buf.write(u"\u0187\7o\2\2\u0187\u0188\7c\2\2\u0188\u0189\7v\2\2\u0189")
+                buf.write(u"\u018a\7j\2\2\u018a\u018b\7k\2\2\u018b\u018c\7v\2\2\u018c")
+                buf.write(u"Z\3\2\2\2\u018d\u018e\7a\2\2\u018e\\\3\2\2\2\u018f\u0190")
+                buf.write(u"\7`\2\2\u0190^\3\2\2\2\u0191\u0192\7<\2\2\u0192`\3\2")
+                buf.write(u"\2\2\u0193\u0194\t\2\2\2\u0194b\3\2\2\2\u0195\u0199\7")
+                buf.write(u"f\2\2\u0196\u0198\5a\61\2\u0197\u0196\3\2\2\2\u0198\u019b")
+                buf.write(u"\3\2\2\2\u0199\u019a\3\2\2\2\u0199\u0197\3\2\2\2\u019a")
+                buf.write(u"\u01a3\3\2\2\2\u019b\u0199\3\2\2\2\u019c\u01a4\t\3\2")
+                buf.write(u"\2\u019d\u019f\7^\2\2\u019e\u01a0\t\3\2\2\u019f\u019e")
+                buf.write(u"\3\2\2\2\u01a0\u01a1\3\2\2\2\u01a1\u019f\3\2\2\2\u01a1")
+                buf.write(u"\u01a2\3\2\2\2\u01a2\u01a4\3\2\2\2\u01a3\u019c\3\2\2")
+                buf.write(u"\2\u01a3\u019d\3\2\2\2\u01a4d\3\2\2\2\u01a5\u01a6\t\3")
+                buf.write(u"\2\2\u01a6f\3\2\2\2\u01a7\u01a8\t\4\2\2\u01a8h\3\2\2")
+                buf.write(u"\2\u01a9\u01ab\5g\64\2\u01aa\u01a9\3\2\2\2\u01ab\u01ac")
+                buf.write(u"\3\2\2\2\u01ac\u01aa\3\2\2\2\u01ac\u01ad\3\2\2\2\u01ad")
+                buf.write(u"\u01b5\3\2\2\2\u01ae\u01af\7.\2\2\u01af\u01b0\5g\64\2")
+                buf.write(u"\u01b0\u01b1\5g\64\2\u01b1\u01b2\5g\64\2\u01b2\u01b4")
+                buf.write(u"\3\2\2\2\u01b3\u01ae\3\2\2\2\u01b4\u01b7\3\2\2\2\u01b5")
+                buf.write(u"\u01b3\3\2\2\2\u01b5\u01b6\3\2\2\2\u01b6\u01cf\3\2\2")
+                buf.write(u"\2\u01b7\u01b5\3\2\2\2\u01b8\u01ba\5g\64\2\u01b9\u01b8")
+                buf.write(u"\3\2\2\2\u01ba\u01bd\3\2\2\2\u01bb\u01b9\3\2\2\2\u01bb")
+                buf.write(u"\u01bc\3\2\2\2\u01bc\u01c5\3\2\2\2\u01bd\u01bb\3\2\2")
+                buf.write(u"\2\u01be\u01bf\7.\2\2\u01bf\u01c0\5g\64\2\u01c0\u01c1")
+                buf.write(u"\5g\64\2\u01c1\u01c2\5g\64\2\u01c2\u01c4\3\2\2\2\u01c3")
+                buf.write(u"\u01be\3\2\2\2\u01c4\u01c7\3\2\2\2\u01c5\u01c3\3\2\2")
+                buf.write(u"\2\u01c5\u01c6\3\2\2\2\u01c6\u01c8\3\2\2\2\u01c7\u01c5")
+                buf.write(u"\3\2\2\2\u01c8\u01ca\7\60\2\2\u01c9\u01cb\5g\64\2\u01ca")
+                buf.write(u"\u01c9\3\2\2\2\u01cb\u01cc\3\2\2\2\u01cc\u01ca\3\2\2")
+                buf.write(u"\2\u01cc\u01cd\3\2\2\2\u01cd\u01cf\3\2\2\2\u01ce\u01aa")
+                buf.write(u"\3\2\2\2\u01ce\u01bb\3\2\2\2\u01cfj\3\2\2\2\u01d0\u01d1")
+                buf.write(u"\7?\2\2\u01d1l\3\2\2\2\u01d2\u01d3\7>\2\2\u01d3n\3\2")
+                buf.write(u"\2\2\u01d4\u01d5\7^\2\2\u01d5\u01d6\7n\2\2\u01d6\u01d7")
+                buf.write(u"\7g\2\2\u01d7\u01d8\7s\2\2\u01d8p\3\2\2\2\u01d9\u01da")
+                buf.write(u"\7@\2\2\u01dar\3\2\2\2\u01db\u01dc\7^\2\2\u01dc\u01dd")
+                buf.write(u"\7i\2\2\u01dd\u01de\7g\2\2\u01de\u01df\7s\2\2\u01dft")
+                buf.write(u"\3\2\2\2\u01e0\u01e1\7#\2\2\u01e1v\3\2\2\2\u01e2\u01e4")
+                buf.write(u"\7^\2\2\u01e3\u01e5\t\3\2\2\u01e4\u01e3\3\2\2\2\u01e5")
+                buf.write(u"\u01e6\3\2\2\2\u01e6\u01e4\3\2\2\2\u01e6\u01e7\3\2\2")
+                buf.write(u"\2\u01e7x\3\2\2\2\17\2~\u00d4\u0199\u01a1\u01a3\u01ac")
+                buf.write(u"\u01b5\u01bb\u01c5\u01cc\u01ce\u01e6\3\b\2\2")
+                return buf.getvalue()
+        
+        
+        class LaTeXLexer(Lexer):
+        
+            atn = ATNDeserializer().deserialize(serializedATN())
+        
+            decisionsToDFA = [ DFA(ds, i) for i, ds in enumerate(atn.decisionToState) ]
+        
+            T__0 = 1
+            WS = 2
+            ADD = 3
+            SUB = 4
+            MUL = 5
+            DIV = 6
+            L_PAREN = 7
+            R_PAREN = 8
+            L_BRACE = 9
+            R_BRACE = 10
+            L_BRACKET = 11
+            R_BRACKET = 12
+            BAR = 13
+            FUNC_LIM = 14
+            LIM_APPROACH_SYM = 15
+            FUNC_INT = 16
+            FUNC_SUM = 17
+            FUNC_PROD = 18
+            FUNC_LOG = 19
+            FUNC_LN = 20
+            FUNC_SIN = 21
+            FUNC_COS = 22
+            FUNC_TAN = 23
+            FUNC_CSC = 24
+            FUNC_SEC = 25
+            FUNC_COT = 26
+            FUNC_ARCSIN = 27
+            FUNC_ARCCOS = 28
+            FUNC_ARCTAN = 29
+            FUNC_ARCCSC = 30
+            FUNC_ARCSEC = 31
+            FUNC_ARCCOT = 32
+            FUNC_SINH = 33
+            FUNC_COSH = 34
+            FUNC_TANH = 35
+            FUNC_ARSINH = 36
+            FUNC_ARCOSH = 37
+            FUNC_ARTANH = 38
+            FUNC_SQRT = 39
+            CMD_TIMES = 40
+            CMD_CDOT = 41
+            CMD_DIV = 42
+            CMD_FRAC = 43
+            CMD_MATHIT = 44
+            UNDERSCORE = 45
+            CARET = 46
+            COLON = 47
+            DIFFERENTIAL = 48
+            LETTER = 49
+            NUMBER = 50
+            EQUAL = 51
+            LT = 52
+            LTE = 53
+            GT = 54
+            GTE = 55
+            BANG = 56
+            SYMBOL = 57
+        
+            channelNames = [ u"DEFAULT_TOKEN_CHANNEL", u"HIDDEN" ]
+        
+            modeNames = [ u"DEFAULT_MODE" ]
+        
+            literalNames = [ u"<INVALID>",
+                    u"','", u"'+'", u"'-'", u"'*'", u"'/'", u"'('", u"')'", u"'{'", 
+                    u"'}'", u"'['", u"']'", u"'|'", u"'\\lim'", u"'\\int'", u"'\\sum'", 
+                    u"'\\prod'", u"'\\log'", u"'\\ln'", u"'\\sin'", u"'\\cos'", 
+                    u"'\\tan'", u"'\\csc'", u"'\\sec'", u"'\\cot'", u"'\\arcsin'", 
+                    u"'\\arccos'", u"'\\arctan'", u"'\\arccsc'", u"'\\arcsec'", 
+                    u"'\\arccot'", u"'\\sinh'", u"'\\cosh'", u"'\\tanh'", u"'\\arsinh'", 
+                    u"'\\arcosh'", u"'\\artanh'", u"'\\sqrt'", u"'\\times'", u"'\\cdot'", 
+                    u"'\\div'", u"'\\frac'", u"'\\mathit'", u"'_'", u"'^'", u"':'", 
+                    u"'='", u"'<'", u"'\\leq'", u"'>'", u"'\\geq'", u"'!'" ]
+        
+            symbolicNames = [ u"<INVALID>",
+                    u"WS", u"ADD", u"SUB", u"MUL", u"DIV", u"L_PAREN", u"R_PAREN", 
+                    u"L_BRACE", u"R_BRACE", u"L_BRACKET", u"R_BRACKET", u"BAR", 
+                    u"FUNC_LIM", u"LIM_APPROACH_SYM", u"FUNC_INT", u"FUNC_SUM", 
+                    u"FUNC_PROD", u"FUNC_LOG", u"FUNC_LN", u"FUNC_SIN", u"FUNC_COS", 
+                    u"FUNC_TAN", u"FUNC_CSC", u"FUNC_SEC", u"FUNC_COT", u"FUNC_ARCSIN", 
+                    u"FUNC_ARCCOS", u"FUNC_ARCTAN", u"FUNC_ARCCSC", u"FUNC_ARCSEC", 
+                    u"FUNC_ARCCOT", u"FUNC_SINH", u"FUNC_COSH", u"FUNC_TANH", u"FUNC_ARSINH", 
+                    u"FUNC_ARCOSH", u"FUNC_ARTANH", u"FUNC_SQRT", u"CMD_TIMES", 
+                    u"CMD_CDOT", u"CMD_DIV", u"CMD_FRAC", u"CMD_MATHIT", u"UNDERSCORE", 
+                    u"CARET", u"COLON", u"DIFFERENTIAL", u"LETTER", u"NUMBER", u"EQUAL", 
+                    u"LT", u"LTE", u"GT", u"GTE", u"BANG", u"SYMBOL" ]
+        
+            ruleNames = [ u"T__0", u"WS", u"ADD", u"SUB", u"MUL", u"DIV", u"L_PAREN", 
+                          u"R_PAREN", u"L_BRACE", u"R_BRACE", u"L_BRACKET", u"R_BRACKET", 
+                          u"BAR", u"FUNC_LIM", u"LIM_APPROACH_SYM", u"FUNC_INT", 
+                          u"FUNC_SUM", u"FUNC_PROD", u"FUNC_LOG", u"FUNC_LN", u"FUNC_SIN", 
+                          u"FUNC_COS", u"FUNC_TAN", u"FUNC_CSC", u"FUNC_SEC", u"FUNC_COT", 
+                          u"FUNC_ARCSIN", u"FUNC_ARCCOS", u"FUNC_ARCTAN", u"FUNC_ARCCSC", 
+                          u"FUNC_ARCSEC", u"FUNC_ARCCOT", u"FUNC_SINH", u"FUNC_COSH", 
+                          u"FUNC_TANH", u"FUNC_ARSINH", u"FUNC_ARCOSH", u"FUNC_ARTANH", 
+                          u"FUNC_SQRT", u"CMD_TIMES", u"CMD_CDOT", u"CMD_DIV", u"CMD_FRAC", 
+                          u"CMD_MATHIT", u"UNDERSCORE", u"CARET", u"COLON", u"WS_CHAR", 
+                          u"DIFFERENTIAL", u"LETTER", u"DIGIT", u"NUMBER", u"EQUAL", 
+                          u"LT", u"LTE", u"GT", u"GTE", u"BANG", u"SYMBOL" ]
+        
+            grammarFileName = u"LaTeX.g4"
+        
+            def __init__(self, input=None, output=sys.stdout):
+                super(LaTeXLexer, self).__init__(input, output=output)
+                self.checkVersion("4.7.1")
+                self._interp = LexerATNSimulator(self, self.atn, self.decisionsToDFA, PredictionContextCache())
+                self._actions = None
+                self._predicates = None
+        
+        
+    except Exception as err:
+        print(err)
+    return locals()
