--- sympy/parsing/_latex/latexlexer.tokens
@@ -1,108 +1,0 @@
-T__0=1
-WS=2
-ADD=3
-SUB=4
-MUL=5
-DIV=6
-L_PAREN=7
-R_PAREN=8
-L_BRACE=9
-R_BRACE=10
-L_BRACKET=11
-R_BRACKET=12
-BAR=13
-FUNC_LIM=14
-LIM_APPROACH_SYM=15
-FUNC_INT=16
-FUNC_SUM=17
-FUNC_PROD=18
-FUNC_LOG=19
-FUNC_LN=20
-FUNC_SIN=21
-FUNC_COS=22
-FUNC_TAN=23
-FUNC_CSC=24
-FUNC_SEC=25
-FUNC_COT=26
-FUNC_ARCSIN=27
-FUNC_ARCCOS=28
-FUNC_ARCTAN=29
-FUNC_ARCCSC=30
-FUNC_ARCSEC=31
-FUNC_ARCCOT=32
-FUNC_SINH=33
-FUNC_COSH=34
-FUNC_TANH=35
-FUNC_ARSINH=36
-FUNC_ARCOSH=37
-FUNC_ARTANH=38
-FUNC_SQRT=39
-CMD_TIMES=40
-CMD_CDOT=41
-CMD_DIV=42
-CMD_FRAC=43
-CMD_MATHIT=44
-UNDERSCORE=45
-CARET=46
-COLON=47
-DIFFERENTIAL=48
-LETTER=49
-NUMBER=50
-EQUAL=51
-LT=52
-LTE=53
-GT=54
-GTE=55
-BANG=56
-SYMBOL=57
-','=1
-'+'=3
-'-'=4
-'*'=5
-'/'=6
-'('=7
-')'=8
-'{'=9
-'}'=10
-'['=11
-']'=12
-'|'=13
-'\\lim'=14
-'\\int'=16
-'\\sum'=17
-'\\prod'=18
-'\\log'=19
-'\\ln'=20
-'\\sin'=21
-'\\cos'=22
-'\\tan'=23
-'\\csc'=24
-'\\sec'=25
-'\\cot'=26
-'\\arcsin'=27
-'\\arccos'=28
-'\\arctan'=29
-'\\arccsc'=30
-'\\arcsec'=31
-'\\arccot'=32
-'\\sinh'=33
-'\\cosh'=34
-'\\tanh'=35
-'\\arsinh'=36
-'\\arcosh'=37
-'\\artanh'=38
-'\\sqrt'=39
-'\\times'=40
-'\\cdot'=41
-'\\div'=42
-'\\frac'=43
-'\\mathit'=44
-'_'=45
-'^'=46
-':'=47
-'='=51
-'<'=52
-'\\leq'=53
-'>'=54
-'\\geq'=55
-'!'=56
