--- sympy/parsing/_latex/latexlistener.py
@@ -1,321 +1,0 @@
-# Generated from LaTeX.g4 by ANTLR 4.7
-from antlr4 import *
-
-# This class defines a complete listener for a parse tree produced by LaTeXParser.
-class LaTeXListener(ParseTreeListener):
-
-    # Enter a parse tree produced by LaTeXParser#math.
-    def enterMath(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#math.
-    def exitMath(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#relation.
-    def enterRelation(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#relation.
-    def exitRelation(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#equality.
-    def enterEquality(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#equality.
-    def exitEquality(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#expr.
-    def enterExpr(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#expr.
-    def exitExpr(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#additive.
-    def enterAdditive(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#additive.
-    def exitAdditive(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#mp.
-    def enterMp(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#mp.
-    def exitMp(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#mp_nofunc.
-    def enterMp_nofunc(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#mp_nofunc.
-    def exitMp_nofunc(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#unary.
-    def enterUnary(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#unary.
-    def exitUnary(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#unary_nofunc.
-    def enterUnary_nofunc(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#unary_nofunc.
-    def exitUnary_nofunc(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#postfix.
-    def enterPostfix(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#postfix.
-    def exitPostfix(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#postfix_nofunc.
-    def enterPostfix_nofunc(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#postfix_nofunc.
-    def exitPostfix_nofunc(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#postfix_op.
-    def enterPostfix_op(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#postfix_op.
-    def exitPostfix_op(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#eval_at.
-    def enterEval_at(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#eval_at.
-    def exitEval_at(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#eval_at_sub.
-    def enterEval_at_sub(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#eval_at_sub.
-    def exitEval_at_sub(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#eval_at_sup.
-    def enterEval_at_sup(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#eval_at_sup.
-    def exitEval_at_sup(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#exp.
-    def enterExp(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#exp.
-    def exitExp(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#exp_nofunc.
-    def enterExp_nofunc(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#exp_nofunc.
-    def exitExp_nofunc(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#comp.
-    def enterComp(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#comp.
-    def exitComp(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#comp_nofunc.
-    def enterComp_nofunc(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#comp_nofunc.
-    def exitComp_nofunc(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#group.
-    def enterGroup(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#group.
-    def exitGroup(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#abs_group.
-    def enterAbs_group(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#abs_group.
-    def exitAbs_group(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#atom.
-    def enterAtom(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#atom.
-    def exitAtom(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#mathit.
-    def enterMathit(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#mathit.
-    def exitMathit(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#mathit_text.
-    def enterMathit_text(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#mathit_text.
-    def exitMathit_text(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#frac.
-    def enterFrac(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#frac.
-    def exitFrac(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#func_normal.
-    def enterFunc_normal(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#func_normal.
-    def exitFunc_normal(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#func.
-    def enterFunc(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#func.
-    def exitFunc(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#args.
-    def enterArgs(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#args.
-    def exitArgs(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#limit_sub.
-    def enterLimit_sub(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#limit_sub.
-    def exitLimit_sub(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#func_arg.
-    def enterFunc_arg(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#func_arg.
-    def exitFunc_arg(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#func_arg_noparens.
-    def enterFunc_arg_noparens(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#func_arg_noparens.
-    def exitFunc_arg_noparens(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#subexpr.
-    def enterSubexpr(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#subexpr.
-    def exitSubexpr(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#supexpr.
-    def enterSupexpr(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#supexpr.
-    def exitSupexpr(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#subeq.
-    def enterSubeq(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#subeq.
-    def exitSubeq(self, ctx):
-        pass
-
-
-    # Enter a parse tree produced by LaTeXParser#supeq.
-    def enterSupeq(self, ctx):
-        pass
-
-    # Exit a parse tree produced by LaTeXParser#supeq.
-    def exitSupeq(self, ctx):
-        pass
-
-
