--- sympy/parsing/_latex/latexparser.py
@@ -1,3226 +1,0 @@
-# Generated from LaTeX.g4 by ANTLR 4.7
-# encoding: utf-8
-from __future__ import print_function
-from antlr4 import *
-from io import StringIO
-import sys
-
-def serializedATN():
-    with StringIO() as buf:
-        buf.write(u"\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\3")
-        buf.write(u";\u01a2\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t")
-        buf.write(u"\7\4\b\t\b\4\t\t\t\4\n\t\n\4\13\t\13\4\f\t\f\4\r\t\r")
-        buf.write(u"\4\16\t\16\4\17\t\17\4\20\t\20\4\21\t\21\4\22\t\22\4")
-        buf.write(u"\23\t\23\4\24\t\24\4\25\t\25\4\26\t\26\4\27\t\27\4\30")
-        buf.write(u"\t\30\4\31\t\31\4\32\t\32\4\33\t\33\4\34\t\34\4\35\t")
-        buf.write(u"\35\4\36\t\36\4\37\t\37\4 \t \4!\t!\4\"\t\"\4#\t#\4$")
-        buf.write(u"\t$\3\2\3\2\3\3\3\3\3\3\3\3\3\3\3\3\7\3Q\n\3\f\3\16\3")
-        buf.write(u"T\13\3\3\4\3\4\3\4\3\4\3\5\3\5\3\6\3\6\3\6\3\6\3\6\3")
-        buf.write(u"\6\7\6b\n\6\f\6\16\6e\13\6\3\7\3\7\3\7\3\7\3\7\3\7\7")
-        buf.write(u"\7m\n\7\f\7\16\7p\13\7\3\b\3\b\3\b\3\b\3\b\3\b\7\bx\n")
-        buf.write(u"\b\f\b\16\b{\13\b\3\t\3\t\3\t\6\t\u0080\n\t\r\t\16\t")
-        buf.write(u"\u0081\5\t\u0084\n\t\3\n\3\n\3\n\3\n\7\n\u008a\n\n\f")
-        buf.write(u"\n\16\n\u008d\13\n\5\n\u008f\n\n\3\13\3\13\7\13\u0093")
-        buf.write(u"\n\13\f\13\16\13\u0096\13\13\3\f\3\f\7\f\u009a\n\f\f")
-        buf.write(u"\f\16\f\u009d\13\f\3\r\3\r\5\r\u00a1\n\r\3\16\3\16\3")
-        buf.write(u"\16\3\16\3\16\3\16\5\16\u00a9\n\16\3\17\3\17\3\17\3\17")
-        buf.write(u"\5\17\u00af\n\17\3\17\3\17\3\20\3\20\3\20\3\20\5\20\u00b7")
-        buf.write(u"\n\20\3\20\3\20\3\21\3\21\3\21\3\21\3\21\3\21\3\21\3")
-        buf.write(u"\21\3\21\3\21\5\21\u00c5\n\21\3\21\5\21\u00c8\n\21\7")
-        buf.write(u"\21\u00ca\n\21\f\21\16\21\u00cd\13\21\3\22\3\22\3\22")
-        buf.write(u"\3\22\3\22\3\22\3\22\3\22\3\22\3\22\5\22\u00d9\n\22\3")
-        buf.write(u"\22\5\22\u00dc\n\22\7\22\u00de\n\22\f\22\16\22\u00e1")
-        buf.write(u"\13\22\3\23\3\23\3\23\3\23\3\23\5\23\u00e8\n\23\3\24")
-        buf.write(u"\3\24\3\24\3\24\5\24\u00ee\n\24\3\25\3\25\3\25\3\25\3")
-        buf.write(u"\25\3\25\3\25\3\25\3\25\3\25\3\25\3\25\5\25\u00fc\n\25")
-        buf.write(u"\3\26\3\26\3\26\3\26\3\27\3\27\5\27\u0104\n\27\3\27\3")
-        buf.write(u"\27\3\27\5\27\u0109\n\27\3\30\3\30\3\30\3\30\3\30\3\31")
-        buf.write(u"\7\31\u0111\n\31\f\31\16\31\u0114\13\31\3\32\3\32\3\32")
-        buf.write(u"\3\32\3\32\3\32\3\32\3\32\3\33\3\33\3\34\3\34\5\34\u0122")
-        buf.write(u"\n\34\3\34\5\34\u0125\n\34\3\34\5\34\u0128\n\34\3\34")
-        buf.write(u"\5\34\u012b\n\34\5\34\u012d\n\34\3\34\3\34\3\34\3\34")
-        buf.write(u"\3\34\5\34\u0134\n\34\3\34\3\34\5\34\u0138\n\34\3\34")
-        buf.write(u"\3\34\3\34\3\34\3\34\3\34\3\34\3\34\3\34\3\34\3\34\5")
-        buf.write(u"\34\u0145\n\34\3\34\5\34\u0148\n\34\3\34\3\34\3\34\5")
-        buf.write(u"\34\u014d\n\34\3\34\3\34\3\34\3\34\3\34\5\34\u0154\n")
-        buf.write(u"\34\3\34\3\34\3\34\3\34\3\34\3\34\3\34\3\34\3\34\3\34")
-        buf.write(u"\3\34\5\34\u0161\n\34\3\34\3\34\3\34\3\34\3\34\3\34\5")
-        buf.write(u"\34\u0169\n\34\3\35\3\35\3\35\3\35\3\35\5\35\u0170\n")
-        buf.write(u"\35\3\36\3\36\3\36\3\36\3\36\3\36\3\36\3\36\3\36\5\36")
-        buf.write(u"\u017b\n\36\3\36\3\36\3\37\3\37\3\37\3\37\3\37\5\37\u0184")
-        buf.write(u"\n\37\3 \3 \3!\3!\3!\3!\3!\3!\5!\u018e\n!\3\"\3\"\3\"")
-        buf.write(u"\3\"\3\"\3\"\5\"\u0196\n\"\3#\3#\3#\3#\3#\3$\3$\3$\3")
-        buf.write(u"$\3$\3$\2\b\4\n\f\16 \"%\2\4\6\b\n\f\16\20\22\24\26\30")
-        buf.write(u"\32\34\36 \"$&(*,.\60\62\64\668:<>@BDF\2\b\3\2\659\3")
-        buf.write(u"\2\5\6\5\2\7\b*,\61\61\4\2\63\63;;\3\2\25(\3\2\23\24")
-        buf.write(u"\2\u01b9\2H\3\2\2\2\4J\3\2\2\2\6U\3\2\2\2\bY\3\2\2\2")
-        buf.write(u"\n[\3\2\2\2\ff\3\2\2\2\16q\3\2\2\2\20\u0083\3\2\2\2\22")
-        buf.write(u"\u008e\3\2\2\2\24\u0090\3\2\2\2\26\u0097\3\2\2\2\30\u00a0")
-        buf.write(u"\3\2\2\2\32\u00a2\3\2\2\2\34\u00aa\3\2\2\2\36\u00b2\3")
-        buf.write(u"\2\2\2 \u00ba\3\2\2\2\"\u00ce\3\2\2\2$\u00e7\3\2\2\2")
-        buf.write(u"&\u00ed\3\2\2\2(\u00fb\3\2\2\2*\u00fd\3\2\2\2,\u0108")
-        buf.write(u"\3\2\2\2.\u010a\3\2\2\2\60\u0112\3\2\2\2\62\u0115\3\2")
-        buf.write(u"\2\2\64\u011d\3\2\2\2\66\u0168\3\2\2\28\u016f\3\2\2\2")
-        buf.write(u":\u0171\3\2\2\2<\u0183\3\2\2\2>\u0185\3\2\2\2@\u0187")
-        buf.write(u"\3\2\2\2B\u018f\3\2\2\2D\u0197\3\2\2\2F\u019c\3\2\2\2")
-        buf.write(u"HI\5\4\3\2I\3\3\2\2\2JK\b\3\1\2KL\5\b\5\2LR\3\2\2\2M")
-        buf.write(u"N\f\4\2\2NO\t\2\2\2OQ\5\4\3\5PM\3\2\2\2QT\3\2\2\2RP\3")
-        buf.write(u"\2\2\2RS\3\2\2\2S\5\3\2\2\2TR\3\2\2\2UV\5\b\5\2VW\7\65")
-        buf.write(u"\2\2WX\5\b\5\2X\7\3\2\2\2YZ\5\n\6\2Z\t\3\2\2\2[\\\b\6")
-        buf.write(u"\1\2\\]\5\f\7\2]c\3\2\2\2^_\f\4\2\2_`\t\3\2\2`b\5\n\6")
-        buf.write(u"\5a^\3\2\2\2be\3\2\2\2ca\3\2\2\2cd\3\2\2\2d\13\3\2\2")
-        buf.write(u"\2ec\3\2\2\2fg\b\7\1\2gh\5\20\t\2hn\3\2\2\2ij\f\4\2\2")
-        buf.write(u"jk\t\4\2\2km\5\f\7\5li\3\2\2\2mp\3\2\2\2nl\3\2\2\2no")
-        buf.write(u"\3\2\2\2o\r\3\2\2\2pn\3\2\2\2qr\b\b\1\2rs\5\22\n\2sy")
-        buf.write(u"\3\2\2\2tu\f\4\2\2uv\t\4\2\2vx\5\16\b\5wt\3\2\2\2x{\3")
-        buf.write(u"\2\2\2yw\3\2\2\2yz\3\2\2\2z\17\3\2\2\2{y\3\2\2\2|}\t")
-        buf.write(u"\3\2\2}\u0084\5\20\t\2~\u0080\5\24\13\2\177~\3\2\2\2")
-        buf.write(u"\u0080\u0081\3\2\2\2\u0081\177\3\2\2\2\u0081\u0082\3")
-        buf.write(u"\2\2\2\u0082\u0084\3\2\2\2\u0083|\3\2\2\2\u0083\177\3")
-        buf.write(u"\2\2\2\u0084\21\3\2\2\2\u0085\u0086\t\3\2\2\u0086\u008f")
-        buf.write(u"\5\22\n\2\u0087\u008b\5\24\13\2\u0088\u008a\5\26\f\2")
-        buf.write(u"\u0089\u0088\3\2\2\2\u008a\u008d\3\2\2\2\u008b\u0089")
-        buf.write(u"\3\2\2\2\u008b\u008c\3\2\2\2\u008c\u008f\3\2\2\2\u008d")
-        buf.write(u"\u008b\3\2\2\2\u008e\u0085\3\2\2\2\u008e\u0087\3\2\2")
-        buf.write(u"\2\u008f\23\3\2\2\2\u0090\u0094\5 \21\2\u0091\u0093\5")
-        buf.write(u"\30\r\2\u0092\u0091\3\2\2\2\u0093\u0096\3\2\2\2\u0094")
-        buf.write(u"\u0092\3\2\2\2\u0094\u0095\3\2\2\2\u0095\25\3\2\2\2\u0096")
-        buf.write(u"\u0094\3\2\2\2\u0097\u009b\5\"\22\2\u0098\u009a\5\30")
-        buf.write(u"\r\2\u0099\u0098\3\2\2\2\u009a\u009d\3\2\2\2\u009b\u0099")
-        buf.write(u"\3\2\2\2\u009b\u009c\3\2\2\2\u009c\27\3\2\2\2\u009d\u009b")
-        buf.write(u"\3\2\2\2\u009e\u00a1\7:\2\2\u009f\u00a1\5\32\16\2\u00a0")
-        buf.write(u"\u009e\3\2\2\2\u00a0\u009f\3\2\2\2\u00a1\31\3\2\2\2\u00a2")
-        buf.write(u"\u00a8\7\17\2\2\u00a3\u00a9\5\36\20\2\u00a4\u00a9\5\34")
-        buf.write(u"\17\2\u00a5\u00a6\5\36\20\2\u00a6\u00a7\5\34\17\2\u00a7")
-        buf.write(u"\u00a9\3\2\2\2\u00a8\u00a3\3\2\2\2\u00a8\u00a4\3\2\2")
-        buf.write(u"\2\u00a8\u00a5\3\2\2\2\u00a9\33\3\2\2\2\u00aa\u00ab\7")
-        buf.write(u"/\2\2\u00ab\u00ae\7\13\2\2\u00ac\u00af\5\b\5\2\u00ad")
-        buf.write(u"\u00af\5\6\4\2\u00ae\u00ac\3\2\2\2\u00ae\u00ad\3\2\2")
-        buf.write(u"\2\u00af\u00b0\3\2\2\2\u00b0\u00b1\7\f\2\2\u00b1\35\3")
-        buf.write(u"\2\2\2\u00b2\u00b3\7\60\2\2\u00b3\u00b6\7\13\2\2\u00b4")
-        buf.write(u"\u00b7\5\b\5\2\u00b5\u00b7\5\6\4\2\u00b6\u00b4\3\2\2")
-        buf.write(u"\2\u00b6\u00b5\3\2\2\2\u00b7\u00b8\3\2\2\2\u00b8\u00b9")
-        buf.write(u"\7\f\2\2\u00b9\37\3\2\2\2\u00ba\u00bb\b\21\1\2\u00bb")
-        buf.write(u"\u00bc\5$\23\2\u00bc\u00cb\3\2\2\2\u00bd\u00be\f\4\2")
-        buf.write(u"\2\u00be\u00c4\7\60\2\2\u00bf\u00c5\5,\27\2\u00c0\u00c1")
-        buf.write(u"\7\13\2\2\u00c1\u00c2\5\b\5\2\u00c2\u00c3\7\f\2\2\u00c3")
-        buf.write(u"\u00c5\3\2\2\2\u00c4\u00bf\3\2\2\2\u00c4\u00c0\3\2\2")
-        buf.write(u"\2\u00c5\u00c7\3\2\2\2\u00c6\u00c8\5@!\2\u00c7\u00c6")
-        buf.write(u"\3\2\2\2\u00c7\u00c8\3\2\2\2\u00c8\u00ca\3\2\2\2\u00c9")
-        buf.write(u"\u00bd\3\2\2\2\u00ca\u00cd\3\2\2\2\u00cb\u00c9\3\2\2")
-        buf.write(u"\2\u00cb\u00cc\3\2\2\2\u00cc!\3\2\2\2\u00cd\u00cb\3\2")
-        buf.write(u"\2\2\u00ce\u00cf\b\22\1\2\u00cf\u00d0\5&\24\2\u00d0\u00df")
-        buf.write(u"\3\2\2\2\u00d1\u00d2\f\4\2\2\u00d2\u00d8\7\60\2\2\u00d3")
-        buf.write(u"\u00d9\5,\27\2\u00d4\u00d5\7\13\2\2\u00d5\u00d6\5\b\5")
-        buf.write(u"\2\u00d6\u00d7\7\f\2\2\u00d7\u00d9\3\2\2\2\u00d8\u00d3")
-        buf.write(u"\3\2\2\2\u00d8\u00d4\3\2\2\2\u00d9\u00db\3\2\2\2\u00da")
-        buf.write(u"\u00dc\5@!\2\u00db\u00da\3\2\2\2\u00db\u00dc\3\2\2\2")
-        buf.write(u"\u00dc\u00de\3\2\2\2\u00dd\u00d1\3\2\2\2\u00de\u00e1")
-        buf.write(u"\3\2\2\2\u00df\u00dd\3\2\2\2\u00df\u00e0\3\2\2\2\u00e0")
-        buf.write(u"#\3\2\2\2\u00e1\u00df\3\2\2\2\u00e2\u00e8\5(\25\2\u00e3")
-        buf.write(u"\u00e8\5*\26\2\u00e4\u00e8\5\66\34\2\u00e5\u00e8\5,\27")
-        buf.write(u"\2\u00e6\u00e8\5\62\32\2\u00e7\u00e2\3\2\2\2\u00e7\u00e3")
-        buf.write(u"\3\2\2\2\u00e7\u00e4\3\2\2\2\u00e7\u00e5\3\2\2\2\u00e7")
-        buf.write(u"\u00e6\3\2\2\2\u00e8%\3\2\2\2\u00e9\u00ee\5(\25\2\u00ea")
-        buf.write(u"\u00ee\5*\26\2\u00eb\u00ee\5,\27\2\u00ec\u00ee\5\62\32")
-        buf.write(u"\2\u00ed\u00e9\3\2\2\2\u00ed\u00ea\3\2\2\2\u00ed\u00eb")
-        buf.write(u"\3\2\2\2\u00ed\u00ec\3\2\2\2\u00ee\'\3\2\2\2\u00ef\u00f0")
-        buf.write(u"\7\t\2\2\u00f0\u00f1\5\b\5\2\u00f1\u00f2\7\n\2\2\u00f2")
-        buf.write(u"\u00fc\3\2\2\2\u00f3\u00f4\7\r\2\2\u00f4\u00f5\5\b\5")
-        buf.write(u"\2\u00f5\u00f6\7\16\2\2\u00f6\u00fc\3\2\2\2\u00f7\u00f8")
-        buf.write(u"\7\13\2\2\u00f8\u00f9\5\b\5\2\u00f9\u00fa\7\f\2\2\u00fa")
-        buf.write(u"\u00fc\3\2\2\2\u00fb\u00ef\3\2\2\2\u00fb\u00f3\3\2\2")
-        buf.write(u"\2\u00fb\u00f7\3\2\2\2\u00fc)\3\2\2\2\u00fd\u00fe\7\17")
-        buf.write(u"\2\2\u00fe\u00ff\5\b\5\2\u00ff\u0100\7\17\2\2\u0100+")
-        buf.write(u"\3\2\2\2\u0101\u0103\t\5\2\2\u0102\u0104\5@!\2\u0103")
-        buf.write(u"\u0102\3\2\2\2\u0103\u0104\3\2\2\2\u0104\u0109\3\2\2")
-        buf.write(u"\2\u0105\u0109\7\64\2\2\u0106\u0109\7\62\2\2\u0107\u0109")
-        buf.write(u"\5.\30\2\u0108\u0101\3\2\2\2\u0108\u0105\3\2\2\2\u0108")
-        buf.write(u"\u0106\3\2\2\2\u0108\u0107\3\2\2\2\u0109-\3\2\2\2\u010a")
-        buf.write(u"\u010b\7.\2\2\u010b\u010c\7\13\2\2\u010c\u010d\5\60\31")
-        buf.write(u"\2\u010d\u010e\7\f\2\2\u010e/\3\2\2\2\u010f\u0111\7\63")
-        buf.write(u"\2\2\u0110\u010f\3\2\2\2\u0111\u0114\3\2\2\2\u0112\u0110")
-        buf.write(u"\3\2\2\2\u0112\u0113\3\2\2\2\u0113\61\3\2\2\2\u0114\u0112")
-        buf.write(u"\3\2\2\2\u0115\u0116\7-\2\2\u0116\u0117\7\13\2\2\u0117")
-        buf.write(u"\u0118\5\b\5\2\u0118\u0119\7\f\2\2\u0119\u011a\7\13\2")
-        buf.write(u"\2\u011a\u011b\5\b\5\2\u011b\u011c\7\f\2\2\u011c\63\3")
-        buf.write(u"\2\2\2\u011d\u011e\t\6\2\2\u011e\65\3\2\2\2\u011f\u012c")
-        buf.write(u"\5\64\33\2\u0120\u0122\5@!\2\u0121\u0120\3\2\2\2\u0121")
-        buf.write(u"\u0122\3\2\2\2\u0122\u0124\3\2\2\2\u0123\u0125\5B\"\2")
-        buf.write(u"\u0124\u0123\3\2\2\2\u0124\u0125\3\2\2\2\u0125\u012d")
-        buf.write(u"\3\2\2\2\u0126\u0128\5B\"\2\u0127\u0126\3\2\2\2\u0127")
-        buf.write(u"\u0128\3\2\2\2\u0128\u012a\3\2\2\2\u0129\u012b\5@!\2")
-        buf.write(u"\u012a\u0129\3\2\2\2\u012a\u012b\3\2\2\2\u012b\u012d")
-        buf.write(u"\3\2\2\2\u012c\u0121\3\2\2\2\u012c\u0127\3\2\2\2\u012d")
-        buf.write(u"\u0133\3\2\2\2\u012e\u012f\7\t\2\2\u012f\u0130\5<\37")
-        buf.write(u"\2\u0130\u0131\7\n\2\2\u0131\u0134\3\2\2\2\u0132\u0134")
-        buf.write(u"\5> \2\u0133\u012e\3\2\2\2\u0133\u0132\3\2\2\2\u0134")
-        buf.write(u"\u0169\3\2\2\2\u0135\u0137\t\5\2\2\u0136\u0138\5@!\2")
-        buf.write(u"\u0137\u0136\3\2\2\2\u0137\u0138\3\2\2\2\u0138\u0139")
-        buf.write(u"\3\2\2\2\u0139\u013a\7\t\2\2\u013a\u013b\58\35\2\u013b")
-        buf.write(u"\u013c\7\n\2\2\u013c\u0169\3\2\2\2\u013d\u0144\7\22\2")
-        buf.write(u"\2\u013e\u013f\5@!\2\u013f\u0140\5B\"\2\u0140\u0145\3")
-        buf.write(u"\2\2\2\u0141\u0142\5B\"\2\u0142\u0143\5@!\2\u0143\u0145")
-        buf.write(u"\3\2\2\2\u0144\u013e\3\2\2\2\u0144\u0141\3\2\2\2\u0144")
-        buf.write(u"\u0145\3\2\2\2\u0145\u014c\3\2\2\2\u0146\u0148\5\n\6")
-        buf.write(u"\2\u0147\u0146\3\2\2\2\u0147\u0148\3\2\2\2\u0148\u0149")
-        buf.write(u"\3\2\2\2\u0149\u014d\7\62\2\2\u014a\u014d\5\62\32\2\u014b")
-        buf.write(u"\u014d\5\n\6\2\u014c\u0147\3\2\2\2\u014c\u014a\3\2\2")
-        buf.write(u"\2\u014c\u014b\3\2\2\2\u014d\u0169\3\2\2\2\u014e\u0153")
-        buf.write(u"\7)\2\2\u014f\u0150\7\r\2\2\u0150\u0151\5\b\5\2\u0151")
-        buf.write(u"\u0152\7\16\2\2\u0152\u0154\3\2\2\2\u0153\u014f\3\2\2")
-        buf.write(u"\2\u0153\u0154\3\2\2\2\u0154\u0155\3\2\2\2\u0155\u0156")
-        buf.write(u"\7\13\2\2\u0156\u0157\5\b\5\2\u0157\u0158\7\f\2\2\u0158")
-        buf.write(u"\u0169\3\2\2\2\u0159\u0160\t\7\2\2\u015a\u015b\5D#\2")
-        buf.write(u"\u015b\u015c\5B\"\2\u015c\u0161\3\2\2\2\u015d\u015e\5")
-        buf.write(u"B\"\2\u015e\u015f\5D#\2\u015f\u0161\3\2\2\2\u0160\u015a")
-        buf.write(u"\3\2\2\2\u0160\u015d\3\2\2\2\u0161\u0162\3\2\2\2\u0162")
-        buf.write(u"\u0163\5\f\7\2\u0163\u0169\3\2\2\2\u0164\u0165\7\20\2")
-        buf.write(u"\2\u0165\u0166\5:\36\2\u0166\u0167\5\f\7\2\u0167\u0169")
-        buf.write(u"\3\2\2\2\u0168\u011f\3\2\2\2\u0168\u0135\3\2\2\2\u0168")
-        buf.write(u"\u013d\3\2\2\2\u0168\u014e\3\2\2\2\u0168\u0159\3\2\2")
-        buf.write(u"\2\u0168\u0164\3\2\2\2\u0169\67\3\2\2\2\u016a\u016b\5")
-        buf.write(u"\b\5\2\u016b\u016c\7\3\2\2\u016c\u016d\58\35\2\u016d")
-        buf.write(u"\u0170\3\2\2\2\u016e\u0170\5\b\5\2\u016f\u016a\3\2\2")
-        buf.write(u"\2\u016f\u016e\3\2\2\2\u01709\3\2\2\2\u0171\u0172\7/")
-        buf.write(u"\2\2\u0172\u0173\7\13\2\2\u0173\u0174\t\5\2\2\u0174\u0175")
-        buf.write(u"\7\21\2\2\u0175\u017a\5\b\5\2\u0176\u0177\7\60\2\2\u0177")
-        buf.write(u"\u0178\7\13\2\2\u0178\u0179\t\3\2\2\u0179\u017b\7\f\2")
-        buf.write(u"\2\u017a\u0176\3\2\2\2\u017a\u017b\3\2\2\2\u017b\u017c")
-        buf.write(u"\3\2\2\2\u017c\u017d\7\f\2\2\u017d;\3\2\2\2\u017e\u0184")
-        buf.write(u"\5\b\5\2\u017f\u0180\5\b\5\2\u0180\u0181\7\3\2\2\u0181")
-        buf.write(u"\u0182\5<\37\2\u0182\u0184\3\2\2\2\u0183\u017e\3\2\2")
-        buf.write(u"\2\u0183\u017f\3\2\2\2\u0184=\3\2\2\2\u0185\u0186\5\16")
-        buf.write(u"\b\2\u0186?\3\2\2\2\u0187\u018d\7/\2\2\u0188\u018e\5")
-        buf.write(u",\27\2\u0189\u018a\7\13\2\2\u018a\u018b\5\b\5\2\u018b")
-        buf.write(u"\u018c\7\f\2\2\u018c\u018e\3\2\2\2\u018d\u0188\3\2\2")
-        buf.write(u"\2\u018d\u0189\3\2\2\2\u018eA\3\2\2\2\u018f\u0195\7\60")
-        buf.write(u"\2\2\u0190\u0196\5,\27\2\u0191\u0192\7\13\2\2\u0192\u0193")
-        buf.write(u"\5\b\5\2\u0193\u0194\7\f\2\2\u0194\u0196\3\2\2\2\u0195")
-        buf.write(u"\u0190\3\2\2\2\u0195\u0191\3\2\2\2\u0196C\3\2\2\2\u0197")
-        buf.write(u"\u0198\7/\2\2\u0198\u0199\7\13\2\2\u0199\u019a\5\6\4")
-        buf.write(u"\2\u019a\u019b\7\f\2\2\u019bE\3\2\2\2\u019c\u019d\7/")
-        buf.write(u"\2\2\u019d\u019e\7\13\2\2\u019e\u019f\5\6\4\2\u019f\u01a0")
-        buf.write(u"\7\f\2\2\u01a0G\3\2\2\2.Rcny\u0081\u0083\u008b\u008e")
-        buf.write(u"\u0094\u009b\u00a0\u00a8\u00ae\u00b6\u00c4\u00c7\u00cb")
-        buf.write(u"\u00d8\u00db\u00df\u00e7\u00ed\u00fb\u0103\u0108\u0112")
-        buf.write(u"\u0121\u0124\u0127\u012a\u012c\u0133\u0137\u0144\u0147")
-        buf.write(u"\u014c\u0153\u0160\u0168\u016f\u017a\u0183\u018d\u0195")
-        return buf.getvalue()
-
-
-class LaTeXParser ( Parser ):
-
-    grammarFileName = "LaTeX.g4"
-
-    atn = ATNDeserializer().deserialize(serializedATN())
-
-    decisionsToDFA = [ DFA(ds, i) for i, ds in enumerate(atn.decisionToState) ]
-
-    sharedContextCache = PredictionContextCache()
-
-    literalNames = [ u"<INVALID>", u"','", u"<INVALID>", u"'+'", u"'-'", 
-                     u"'*'", u"'/'", u"'('", u"')'", u"'{'", u"'}'", u"'['", 
-                     u"']'", u"'|'", u"'\\lim'", u"<INVALID>", u"'\\int'", 
-                     u"'\\sum'", u"'\\prod'", u"'\\log'", u"'\\ln'", u"'\\sin'", 
-                     u"'\\cos'", u"'\\tan'", u"'\\csc'", u"'\\sec'", u"'\\cot'", 
-                     u"'\\arcsin'", u"'\\arccos'", u"'\\arctan'", u"'\\arccsc'", 
-                     u"'\\arcsec'", u"'\\arccot'", u"'\\sinh'", u"'\\cosh'", 
-                     u"'\\tanh'", u"'\\arsinh'", u"'\\arcosh'", u"'\\artanh'", 
-                     u"'\\sqrt'", u"'\\times'", u"'\\cdot'", u"'\\div'", 
-                     u"'\\frac'", u"'\\mathit'", u"'_'", u"'^'", u"':'", 
-                     u"<INVALID>", u"<INVALID>", u"<INVALID>", u"'='", u"'<'", 
-                     u"'\\leq'", u"'>'", u"'\\geq'", u"'!'" ]
-
-    symbolicNames = [ u"<INVALID>", u"<INVALID>", u"WS", u"ADD", u"SUB", 
-                      u"MUL", u"DIV", u"L_PAREN", u"R_PAREN", u"L_BRACE", 
-                      u"R_BRACE", u"L_BRACKET", u"R_BRACKET", u"BAR", u"FUNC_LIM", 
-                      u"LIM_APPROACH_SYM", u"FUNC_INT", u"FUNC_SUM", u"FUNC_PROD", 
-                      u"FUNC_LOG", u"FUNC_LN", u"FUNC_SIN", u"FUNC_COS", 
-                      u"FUNC_TAN", u"FUNC_CSC", u"FUNC_SEC", u"FUNC_COT", 
-                      u"FUNC_ARCSIN", u"FUNC_ARCCOS", u"FUNC_ARCTAN", u"FUNC_ARCCSC", 
-                      u"FUNC_ARCSEC", u"FUNC_ARCCOT", u"FUNC_SINH", u"FUNC_COSH", 
-                      u"FUNC_TANH", u"FUNC_ARSINH", u"FUNC_ARCOSH", u"FUNC_ARTANH", 
-                      u"FUNC_SQRT", u"CMD_TIMES", u"CMD_CDOT", u"CMD_DIV", 
-                      u"CMD_FRAC", u"CMD_MATHIT", u"UNDERSCORE", u"CARET", 
-                      u"COLON", u"DIFFERENTIAL", u"LETTER", u"NUMBER", u"EQUAL", 
-                      u"LT", u"LTE", u"GT", u"GTE", u"BANG", u"SYMBOL" ]
-
-    RULE_math = 0
-    RULE_relation = 1
-    RULE_equality = 2
-    RULE_expr = 3
-    RULE_additive = 4
-    RULE_mp = 5
-    RULE_mp_nofunc = 6
-    RULE_unary = 7
-    RULE_unary_nofunc = 8
-    RULE_postfix = 9
-    RULE_postfix_nofunc = 10
-    RULE_postfix_op = 11
-    RULE_eval_at = 12
-    RULE_eval_at_sub = 13
-    RULE_eval_at_sup = 14
-    RULE_exp = 15
-    RULE_exp_nofunc = 16
-    RULE_comp = 17
-    RULE_comp_nofunc = 18
-    RULE_group = 19
-    RULE_abs_group = 20
-    RULE_atom = 21
-    RULE_mathit = 22
-    RULE_mathit_text = 23
-    RULE_frac = 24
-    RULE_func_normal = 25
-    RULE_func = 26
-    RULE_args = 27
-    RULE_limit_sub = 28
-    RULE_func_arg = 29
-    RULE_func_arg_noparens = 30
-    RULE_subexpr = 31
-    RULE_supexpr = 32
-    RULE_subeq = 33
-    RULE_supeq = 34
-
-    ruleNames =  [ u"math", u"relation", u"equality", u"expr", u"additive", 
-                   u"mp", u"mp_nofunc", u"unary", u"unary_nofunc", u"postfix", 
-                   u"postfix_nofunc", u"postfix_op", u"eval_at", u"eval_at_sub", 
-                   u"eval_at_sup", u"exp", u"exp_nofunc", u"comp", u"comp_nofunc", 
-                   u"group", u"abs_group", u"atom", u"mathit", u"mathit_text", 
-                   u"frac", u"func_normal", u"func", u"args", u"limit_sub", 
-                   u"func_arg", u"func_arg_noparens", u"subexpr", u"supexpr", 
-                   u"subeq", u"supeq" ]
-
-    EOF = Token.EOF
-    T__0=1
-    WS=2
-    ADD=3
-    SUB=4
-    MUL=5
-    DIV=6
-    L_PAREN=7
-    R_PAREN=8
-    L_BRACE=9
-    R_BRACE=10
-    L_BRACKET=11
-    R_BRACKET=12
-    BAR=13
-    FUNC_LIM=14
-    LIM_APPROACH_SYM=15
-    FUNC_INT=16
-    FUNC_SUM=17
-    FUNC_PROD=18
-    FUNC_LOG=19
-    FUNC_LN=20
-    FUNC_SIN=21
-    FUNC_COS=22
-    FUNC_TAN=23
-    FUNC_CSC=24
-    FUNC_SEC=25
-    FUNC_COT=26
-    FUNC_ARCSIN=27
-    FUNC_ARCCOS=28
-    FUNC_ARCTAN=29
-    FUNC_ARCCSC=30
-    FUNC_ARCSEC=31
-    FUNC_ARCCOT=32
-    FUNC_SINH=33
-    FUNC_COSH=34
-    FUNC_TANH=35
-    FUNC_ARSINH=36
-    FUNC_ARCOSH=37
-    FUNC_ARTANH=38
-    FUNC_SQRT=39
-    CMD_TIMES=40
-    CMD_CDOT=41
-    CMD_DIV=42
-    CMD_FRAC=43
-    CMD_MATHIT=44
-    UNDERSCORE=45
-    CARET=46
-    COLON=47
-    DIFFERENTIAL=48
-    LETTER=49
-    NUMBER=50
-    EQUAL=51
-    LT=52
-    LTE=53
-    GT=54
-    GTE=55
-    BANG=56
-    SYMBOL=57
-
-    def __init__(self, input, output=sys.stdout):
-        super(LaTeXParser, self).__init__(input, output=output)
-        self.checkVersion("4.7")
-        self._interp = ParserATNSimulator(self, self.atn, self.decisionsToDFA, self.sharedContextCache)
-        self._predicates = None
-
-
-
-    class MathContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.MathContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def relation(self):
-            return self.getTypedRuleContext(LaTeXParser.RelationContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_math
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterMath"):
-                listener.enterMath(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitMath"):
-                listener.exitMath(self)
-
-
-
-
-    def math(self):
-
-        localctx = LaTeXParser.MathContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 0, self.RULE_math)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 70
-            self.relation(0)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class RelationContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.RelationContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def relation(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.RelationContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.RelationContext,i)
-
-
-        def EQUAL(self):
-            return self.getToken(LaTeXParser.EQUAL, 0)
-
-        def LT(self):
-            return self.getToken(LaTeXParser.LT, 0)
-
-        def LTE(self):
-            return self.getToken(LaTeXParser.LTE, 0)
-
-        def GT(self):
-            return self.getToken(LaTeXParser.GT, 0)
-
-        def GTE(self):
-            return self.getToken(LaTeXParser.GTE, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_relation
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterRelation"):
-                listener.enterRelation(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitRelation"):
-                listener.exitRelation(self)
-
-
-
-    def relation(self, _p=0):
-        _parentctx = self._ctx
-        _parentState = self.state
-        localctx = LaTeXParser.RelationContext(self, self._ctx, _parentState)
-        _prevctx = localctx
-        _startState = 2
-        self.enterRecursionRule(localctx, 2, self.RULE_relation, _p)
-        self._la = 0 # Token type
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 73
-            self.expr()
-            self._ctx.stop = self._input.LT(-1)
-            self.state = 80
-            self._errHandler.sync(self)
-            _alt = self._interp.adaptivePredict(self._input,0,self._ctx)
-            while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
-                if _alt==1:
-                    if self._parseListeners is not None:
-                        self.triggerExitRuleEvent()
-                    _prevctx = localctx
-                    localctx = LaTeXParser.RelationContext(self, _parentctx, _parentState)
-                    self.pushNewRecursionContext(localctx, _startState, self.RULE_relation)
-                    self.state = 75
-                    if not self.precpred(self._ctx, 2):
-                        from antlr4.error.Errors import FailedPredicateException
-                        raise FailedPredicateException(self, "self.precpred(self._ctx, 2)")
-                    self.state = 76
-                    _la = self._input.LA(1)
-                    if not((((_la) & ~0x3f) == 0 and ((1 << _la) & ((1 << LaTeXParser.EQUAL) | (1 << LaTeXParser.LT) | (1 << LaTeXParser.LTE) | (1 << LaTeXParser.GT) | (1 << LaTeXParser.GTE))) != 0)):
-                        self._errHandler.recoverInline(self)
-                    else:
-                        self._errHandler.reportMatch(self)
-                        self.consume()
-                    self.state = 77
-                    self.relation(3) 
-                self.state = 82
-                self._errHandler.sync(self)
-                _alt = self._interp.adaptivePredict(self._input,0,self._ctx)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.unrollRecursionContexts(_parentctx)
-        return localctx
-
-    class EqualityContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.EqualityContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def expr(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.ExprContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.ExprContext,i)
-
-
-        def EQUAL(self):
-            return self.getToken(LaTeXParser.EQUAL, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_equality
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterEquality"):
-                listener.enterEquality(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitEquality"):
-                listener.exitEquality(self)
-
-
-
-
-    def equality(self):
-
-        localctx = LaTeXParser.EqualityContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 4, self.RULE_equality)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 83
-            self.expr()
-            self.state = 84
-            self.match(LaTeXParser.EQUAL)
-            self.state = 85
-            self.expr()
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class ExprContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.ExprContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def additive(self):
-            return self.getTypedRuleContext(LaTeXParser.AdditiveContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_expr
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterExpr"):
-                listener.enterExpr(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitExpr"):
-                listener.exitExpr(self)
-
-
-
-
-    def expr(self):
-
-        localctx = LaTeXParser.ExprContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 6, self.RULE_expr)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 87
-            self.additive(0)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class AdditiveContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.AdditiveContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def mp(self):
-            return self.getTypedRuleContext(LaTeXParser.MpContext,0)
-
-
-        def additive(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.AdditiveContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.AdditiveContext,i)
-
-
-        def ADD(self):
-            return self.getToken(LaTeXParser.ADD, 0)
-
-        def SUB(self):
-            return self.getToken(LaTeXParser.SUB, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_additive
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterAdditive"):
-                listener.enterAdditive(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitAdditive"):
-                listener.exitAdditive(self)
-
-
-
-    def additive(self, _p=0):
-        _parentctx = self._ctx
-        _parentState = self.state
-        localctx = LaTeXParser.AdditiveContext(self, self._ctx, _parentState)
-        _prevctx = localctx
-        _startState = 8
-        self.enterRecursionRule(localctx, 8, self.RULE_additive, _p)
-        self._la = 0 # Token type
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 90
-            self.mp(0)
-            self._ctx.stop = self._input.LT(-1)
-            self.state = 97
-            self._errHandler.sync(self)
-            _alt = self._interp.adaptivePredict(self._input,1,self._ctx)
-            while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
-                if _alt==1:
-                    if self._parseListeners is not None:
-                        self.triggerExitRuleEvent()
-                    _prevctx = localctx
-                    localctx = LaTeXParser.AdditiveContext(self, _parentctx, _parentState)
-                    self.pushNewRecursionContext(localctx, _startState, self.RULE_additive)
-                    self.state = 92
-                    if not self.precpred(self._ctx, 2):
-                        from antlr4.error.Errors import FailedPredicateException
-                        raise FailedPredicateException(self, "self.precpred(self._ctx, 2)")
-                    self.state = 93
-                    _la = self._input.LA(1)
-                    if not(_la==LaTeXParser.ADD or _la==LaTeXParser.SUB):
-                        self._errHandler.recoverInline(self)
-                    else:
-                        self._errHandler.reportMatch(self)
-                        self.consume()
-                    self.state = 94
-                    self.additive(3) 
-                self.state = 99
-                self._errHandler.sync(self)
-                _alt = self._interp.adaptivePredict(self._input,1,self._ctx)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.unrollRecursionContexts(_parentctx)
-        return localctx
-
-    class MpContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.MpContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def unary(self):
-            return self.getTypedRuleContext(LaTeXParser.UnaryContext,0)
-
-
-        def mp(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.MpContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.MpContext,i)
-
-
-        def MUL(self):
-            return self.getToken(LaTeXParser.MUL, 0)
-
-        def CMD_TIMES(self):
-            return self.getToken(LaTeXParser.CMD_TIMES, 0)
-
-        def CMD_CDOT(self):
-            return self.getToken(LaTeXParser.CMD_CDOT, 0)
-
-        def DIV(self):
-            return self.getToken(LaTeXParser.DIV, 0)
-
-        def CMD_DIV(self):
-            return self.getToken(LaTeXParser.CMD_DIV, 0)
-
-        def COLON(self):
-            return self.getToken(LaTeXParser.COLON, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_mp
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterMp"):
-                listener.enterMp(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitMp"):
-                listener.exitMp(self)
-
-
-
-    def mp(self, _p=0):
-        _parentctx = self._ctx
-        _parentState = self.state
-        localctx = LaTeXParser.MpContext(self, self._ctx, _parentState)
-        _prevctx = localctx
-        _startState = 10
-        self.enterRecursionRule(localctx, 10, self.RULE_mp, _p)
-        self._la = 0 # Token type
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 101
-            self.unary()
-            self._ctx.stop = self._input.LT(-1)
-            self.state = 108
-            self._errHandler.sync(self)
-            _alt = self._interp.adaptivePredict(self._input,2,self._ctx)
-            while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
-                if _alt==1:
-                    if self._parseListeners is not None:
-                        self.triggerExitRuleEvent()
-                    _prevctx = localctx
-                    localctx = LaTeXParser.MpContext(self, _parentctx, _parentState)
-                    self.pushNewRecursionContext(localctx, _startState, self.RULE_mp)
-                    self.state = 103
-                    if not self.precpred(self._ctx, 2):
-                        from antlr4.error.Errors import FailedPredicateException
-                        raise FailedPredicateException(self, "self.precpred(self._ctx, 2)")
-                    self.state = 104
-                    _la = self._input.LA(1)
-                    if not((((_la) & ~0x3f) == 0 and ((1 << _la) & ((1 << LaTeXParser.MUL) | (1 << LaTeXParser.DIV) | (1 << LaTeXParser.CMD_TIMES) | (1 << LaTeXParser.CMD_CDOT) | (1 << LaTeXParser.CMD_DIV) | (1 << LaTeXParser.COLON))) != 0)):
-                        self._errHandler.recoverInline(self)
-                    else:
-                        self._errHandler.reportMatch(self)
-                        self.consume()
-                    self.state = 105
-                    self.mp(3) 
-                self.state = 110
-                self._errHandler.sync(self)
-                _alt = self._interp.adaptivePredict(self._input,2,self._ctx)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.unrollRecursionContexts(_parentctx)
-        return localctx
-
-    class Mp_nofuncContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Mp_nofuncContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def unary_nofunc(self):
-            return self.getTypedRuleContext(LaTeXParser.Unary_nofuncContext,0)
-
-
-        def mp_nofunc(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.Mp_nofuncContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.Mp_nofuncContext,i)
-
-
-        def MUL(self):
-            return self.getToken(LaTeXParser.MUL, 0)
-
-        def CMD_TIMES(self):
-            return self.getToken(LaTeXParser.CMD_TIMES, 0)
-
-        def CMD_CDOT(self):
-            return self.getToken(LaTeXParser.CMD_CDOT, 0)
-
-        def DIV(self):
-            return self.getToken(LaTeXParser.DIV, 0)
-
-        def CMD_DIV(self):
-            return self.getToken(LaTeXParser.CMD_DIV, 0)
-
-        def COLON(self):
-            return self.getToken(LaTeXParser.COLON, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_mp_nofunc
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterMp_nofunc"):
-                listener.enterMp_nofunc(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitMp_nofunc"):
-                listener.exitMp_nofunc(self)
-
-
-
-    def mp_nofunc(self, _p=0):
-        _parentctx = self._ctx
-        _parentState = self.state
-        localctx = LaTeXParser.Mp_nofuncContext(self, self._ctx, _parentState)
-        _prevctx = localctx
-        _startState = 12
-        self.enterRecursionRule(localctx, 12, self.RULE_mp_nofunc, _p)
-        self._la = 0 # Token type
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 112
-            self.unary_nofunc()
-            self._ctx.stop = self._input.LT(-1)
-            self.state = 119
-            self._errHandler.sync(self)
-            _alt = self._interp.adaptivePredict(self._input,3,self._ctx)
-            while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
-                if _alt==1:
-                    if self._parseListeners is not None:
-                        self.triggerExitRuleEvent()
-                    _prevctx = localctx
-                    localctx = LaTeXParser.Mp_nofuncContext(self, _parentctx, _parentState)
-                    self.pushNewRecursionContext(localctx, _startState, self.RULE_mp_nofunc)
-                    self.state = 114
-                    if not self.precpred(self._ctx, 2):
-                        from antlr4.error.Errors import FailedPredicateException
-                        raise FailedPredicateException(self, "self.precpred(self._ctx, 2)")
-                    self.state = 115
-                    _la = self._input.LA(1)
-                    if not((((_la) & ~0x3f) == 0 and ((1 << _la) & ((1 << LaTeXParser.MUL) | (1 << LaTeXParser.DIV) | (1 << LaTeXParser.CMD_TIMES) | (1 << LaTeXParser.CMD_CDOT) | (1 << LaTeXParser.CMD_DIV) | (1 << LaTeXParser.COLON))) != 0)):
-                        self._errHandler.recoverInline(self)
-                    else:
-                        self._errHandler.reportMatch(self)
-                        self.consume()
-                    self.state = 116
-                    self.mp_nofunc(3) 
-                self.state = 121
-                self._errHandler.sync(self)
-                _alt = self._interp.adaptivePredict(self._input,3,self._ctx)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.unrollRecursionContexts(_parentctx)
-        return localctx
-
-    class UnaryContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.UnaryContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def unary(self):
-            return self.getTypedRuleContext(LaTeXParser.UnaryContext,0)
-
-
-        def ADD(self):
-            return self.getToken(LaTeXParser.ADD, 0)
-
-        def SUB(self):
-            return self.getToken(LaTeXParser.SUB, 0)
-
-        def postfix(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.PostfixContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.PostfixContext,i)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_unary
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterUnary"):
-                listener.enterUnary(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitUnary"):
-                listener.exitUnary(self)
-
-
-
-
-    def unary(self):
-
-        localctx = LaTeXParser.UnaryContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 14, self.RULE_unary)
-        self._la = 0 # Token type
-        try:
-            self.state = 129
-            self._errHandler.sync(self)
-            token = self._input.LA(1)
-            if token in [LaTeXParser.ADD, LaTeXParser.SUB]:
-                self.enterOuterAlt(localctx, 1)
-                self.state = 122
-                _la = self._input.LA(1)
-                if not(_la==LaTeXParser.ADD or _la==LaTeXParser.SUB):
-                    self._errHandler.recoverInline(self)
-                else:
-                    self._errHandler.reportMatch(self)
-                    self.consume()
-                self.state = 123
-                self.unary()
-                pass
-            elif token in [LaTeXParser.L_PAREN, LaTeXParser.L_BRACE, LaTeXParser.L_BRACKET, LaTeXParser.BAR, LaTeXParser.FUNC_LIM, LaTeXParser.FUNC_INT, LaTeXParser.FUNC_SUM, LaTeXParser.FUNC_PROD, LaTeXParser.FUNC_LOG, LaTeXParser.FUNC_LN, LaTeXParser.FUNC_SIN, LaTeXParser.FUNC_COS, LaTeXParser.FUNC_TAN, LaTeXParser.FUNC_CSC, LaTeXParser.FUNC_SEC, LaTeXParser.FUNC_COT, LaTeXParser.FUNC_ARCSIN, LaTeXParser.FUNC_ARCCOS, LaTeXParser.FUNC_ARCTAN, LaTeXParser.FUNC_ARCCSC, LaTeXParser.FUNC_ARCSEC, LaTeXParser.FUNC_ARCCOT, LaTeXParser.FUNC_SINH, LaTeXParser.FUNC_COSH, LaTeXParser.FUNC_TANH, LaTeXParser.FUNC_ARSINH, LaTeXParser.FUNC_ARCOSH, LaTeXParser.FUNC_ARTANH, LaTeXParser.FUNC_SQRT, LaTeXParser.CMD_FRAC, LaTeXParser.CMD_MATHIT, LaTeXParser.DIFFERENTIAL, LaTeXParser.LETTER, LaTeXParser.NUMBER, LaTeXParser.SYMBOL]:
-                self.enterOuterAlt(localctx, 2)
-                self.state = 125 
-                self._errHandler.sync(self)
-                _alt = 1
-                while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
-                    if _alt == 1:
-                        self.state = 124
-                        self.postfix()
-
-                    else:
-                        raise NoViableAltException(self)
-                    self.state = 127 
-                    self._errHandler.sync(self)
-                    _alt = self._interp.adaptivePredict(self._input,4,self._ctx)
-
-                pass
-            else:
-                raise NoViableAltException(self)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Unary_nofuncContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Unary_nofuncContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def unary_nofunc(self):
-            return self.getTypedRuleContext(LaTeXParser.Unary_nofuncContext,0)
-
-
-        def ADD(self):
-            return self.getToken(LaTeXParser.ADD, 0)
-
-        def SUB(self):
-            return self.getToken(LaTeXParser.SUB, 0)
-
-        def postfix(self):
-            return self.getTypedRuleContext(LaTeXParser.PostfixContext,0)
-
-
-        def postfix_nofunc(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.Postfix_nofuncContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.Postfix_nofuncContext,i)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_unary_nofunc
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterUnary_nofunc"):
-                listener.enterUnary_nofunc(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitUnary_nofunc"):
-                listener.exitUnary_nofunc(self)
-
-
-
-
-    def unary_nofunc(self):
-
-        localctx = LaTeXParser.Unary_nofuncContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 16, self.RULE_unary_nofunc)
-        self._la = 0 # Token type
-        try:
-            self.state = 140
-            self._errHandler.sync(self)
-            token = self._input.LA(1)
-            if token in [LaTeXParser.ADD, LaTeXParser.SUB]:
-                self.enterOuterAlt(localctx, 1)
-                self.state = 131
-                _la = self._input.LA(1)
-                if not(_la==LaTeXParser.ADD or _la==LaTeXParser.SUB):
-                    self._errHandler.recoverInline(self)
-                else:
-                    self._errHandler.reportMatch(self)
-                    self.consume()
-                self.state = 132
-                self.unary_nofunc()
-                pass
-            elif token in [LaTeXParser.L_PAREN, LaTeXParser.L_BRACE, LaTeXParser.L_BRACKET, LaTeXParser.BAR, LaTeXParser.FUNC_LIM, LaTeXParser.FUNC_INT, LaTeXParser.FUNC_SUM, LaTeXParser.FUNC_PROD, LaTeXParser.FUNC_LOG, LaTeXParser.FUNC_LN, LaTeXParser.FUNC_SIN, LaTeXParser.FUNC_COS, LaTeXParser.FUNC_TAN, LaTeXParser.FUNC_CSC, LaTeXParser.FUNC_SEC, LaTeXParser.FUNC_COT, LaTeXParser.FUNC_ARCSIN, LaTeXParser.FUNC_ARCCOS, LaTeXParser.FUNC_ARCTAN, LaTeXParser.FUNC_ARCCSC, LaTeXParser.FUNC_ARCSEC, LaTeXParser.FUNC_ARCCOT, LaTeXParser.FUNC_SINH, LaTeXParser.FUNC_COSH, LaTeXParser.FUNC_TANH, LaTeXParser.FUNC_ARSINH, LaTeXParser.FUNC_ARCOSH, LaTeXParser.FUNC_ARTANH, LaTeXParser.FUNC_SQRT, LaTeXParser.CMD_FRAC, LaTeXParser.CMD_MATHIT, LaTeXParser.DIFFERENTIAL, LaTeXParser.LETTER, LaTeXParser.NUMBER, LaTeXParser.SYMBOL]:
-                self.enterOuterAlt(localctx, 2)
-                self.state = 133
-                self.postfix()
-                self.state = 137
-                self._errHandler.sync(self)
-                _alt = self._interp.adaptivePredict(self._input,6,self._ctx)
-                while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
-                    if _alt==1:
-                        self.state = 134
-                        self.postfix_nofunc() 
-                    self.state = 139
-                    self._errHandler.sync(self)
-                    _alt = self._interp.adaptivePredict(self._input,6,self._ctx)
-
-                pass
-            else:
-                raise NoViableAltException(self)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class PostfixContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.PostfixContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def exp(self):
-            return self.getTypedRuleContext(LaTeXParser.ExpContext,0)
-
-
-        def postfix_op(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.Postfix_opContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.Postfix_opContext,i)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_postfix
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterPostfix"):
-                listener.enterPostfix(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitPostfix"):
-                listener.exitPostfix(self)
-
-
-
-
-    def postfix(self):
-
-        localctx = LaTeXParser.PostfixContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 18, self.RULE_postfix)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 142
-            self.exp(0)
-            self.state = 146
-            self._errHandler.sync(self)
-            _alt = self._interp.adaptivePredict(self._input,8,self._ctx)
-            while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
-                if _alt==1:
-                    self.state = 143
-                    self.postfix_op() 
-                self.state = 148
-                self._errHandler.sync(self)
-                _alt = self._interp.adaptivePredict(self._input,8,self._ctx)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Postfix_nofuncContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Postfix_nofuncContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def exp_nofunc(self):
-            return self.getTypedRuleContext(LaTeXParser.Exp_nofuncContext,0)
-
-
-        def postfix_op(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.Postfix_opContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.Postfix_opContext,i)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_postfix_nofunc
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterPostfix_nofunc"):
-                listener.enterPostfix_nofunc(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitPostfix_nofunc"):
-                listener.exitPostfix_nofunc(self)
-
-
-
-
-    def postfix_nofunc(self):
-
-        localctx = LaTeXParser.Postfix_nofuncContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 20, self.RULE_postfix_nofunc)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 149
-            self.exp_nofunc(0)
-            self.state = 153
-            self._errHandler.sync(self)
-            _alt = self._interp.adaptivePredict(self._input,9,self._ctx)
-            while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
-                if _alt==1:
-                    self.state = 150
-                    self.postfix_op() 
-                self.state = 155
-                self._errHandler.sync(self)
-                _alt = self._interp.adaptivePredict(self._input,9,self._ctx)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Postfix_opContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Postfix_opContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def BANG(self):
-            return self.getToken(LaTeXParser.BANG, 0)
-
-        def eval_at(self):
-            return self.getTypedRuleContext(LaTeXParser.Eval_atContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_postfix_op
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterPostfix_op"):
-                listener.enterPostfix_op(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitPostfix_op"):
-                listener.exitPostfix_op(self)
-
-
-
-
-    def postfix_op(self):
-
-        localctx = LaTeXParser.Postfix_opContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 22, self.RULE_postfix_op)
-        try:
-            self.state = 158
-            self._errHandler.sync(self)
-            token = self._input.LA(1)
-            if token in [LaTeXParser.BANG]:
-                self.enterOuterAlt(localctx, 1)
-                self.state = 156
-                self.match(LaTeXParser.BANG)
-                pass
-            elif token in [LaTeXParser.BAR]:
-                self.enterOuterAlt(localctx, 2)
-                self.state = 157
-                self.eval_at()
-                pass
-            else:
-                raise NoViableAltException(self)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Eval_atContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Eval_atContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def BAR(self):
-            return self.getToken(LaTeXParser.BAR, 0)
-
-        def eval_at_sup(self):
-            return self.getTypedRuleContext(LaTeXParser.Eval_at_supContext,0)
-
-
-        def eval_at_sub(self):
-            return self.getTypedRuleContext(LaTeXParser.Eval_at_subContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_eval_at
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterEval_at"):
-                listener.enterEval_at(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitEval_at"):
-                listener.exitEval_at(self)
-
-
-
-
-    def eval_at(self):
-
-        localctx = LaTeXParser.Eval_atContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 24, self.RULE_eval_at)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 160
-            self.match(LaTeXParser.BAR)
-            self.state = 166
-            self._errHandler.sync(self)
-            la_ = self._interp.adaptivePredict(self._input,11,self._ctx)
-            if la_ == 1:
-                self.state = 161
-                self.eval_at_sup()
-                pass
-
-            elif la_ == 2:
-                self.state = 162
-                self.eval_at_sub()
-                pass
-
-            elif la_ == 3:
-                self.state = 163
-                self.eval_at_sup()
-                self.state = 164
-                self.eval_at_sub()
-                pass
-
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Eval_at_subContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Eval_at_subContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def UNDERSCORE(self):
-            return self.getToken(LaTeXParser.UNDERSCORE, 0)
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def equality(self):
-            return self.getTypedRuleContext(LaTeXParser.EqualityContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_eval_at_sub
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterEval_at_sub"):
-                listener.enterEval_at_sub(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitEval_at_sub"):
-                listener.exitEval_at_sub(self)
-
-
-
-
-    def eval_at_sub(self):
-
-        localctx = LaTeXParser.Eval_at_subContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 26, self.RULE_eval_at_sub)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 168
-            self.match(LaTeXParser.UNDERSCORE)
-            self.state = 169
-            self.match(LaTeXParser.L_BRACE)
-            self.state = 172
-            self._errHandler.sync(self)
-            la_ = self._interp.adaptivePredict(self._input,12,self._ctx)
-            if la_ == 1:
-                self.state = 170
-                self.expr()
-                pass
-
-            elif la_ == 2:
-                self.state = 171
-                self.equality()
-                pass
-
-
-            self.state = 174
-            self.match(LaTeXParser.R_BRACE)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Eval_at_supContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Eval_at_supContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def CARET(self):
-            return self.getToken(LaTeXParser.CARET, 0)
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def equality(self):
-            return self.getTypedRuleContext(LaTeXParser.EqualityContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_eval_at_sup
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterEval_at_sup"):
-                listener.enterEval_at_sup(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitEval_at_sup"):
-                listener.exitEval_at_sup(self)
-
-
-
-
-    def eval_at_sup(self):
-
-        localctx = LaTeXParser.Eval_at_supContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 28, self.RULE_eval_at_sup)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 176
-            self.match(LaTeXParser.CARET)
-            self.state = 177
-            self.match(LaTeXParser.L_BRACE)
-            self.state = 180
-            self._errHandler.sync(self)
-            la_ = self._interp.adaptivePredict(self._input,13,self._ctx)
-            if la_ == 1:
-                self.state = 178
-                self.expr()
-                pass
-
-            elif la_ == 2:
-                self.state = 179
-                self.equality()
-                pass
-
-
-            self.state = 182
-            self.match(LaTeXParser.R_BRACE)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class ExpContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.ExpContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def comp(self):
-            return self.getTypedRuleContext(LaTeXParser.CompContext,0)
-
-
-        def exp(self):
-            return self.getTypedRuleContext(LaTeXParser.ExpContext,0)
-
-
-        def CARET(self):
-            return self.getToken(LaTeXParser.CARET, 0)
-
-        def atom(self):
-            return self.getTypedRuleContext(LaTeXParser.AtomContext,0)
-
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def subexpr(self):
-            return self.getTypedRuleContext(LaTeXParser.SubexprContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_exp
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterExp"):
-                listener.enterExp(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitExp"):
-                listener.exitExp(self)
-
-
-
-    def exp(self, _p=0):
-        _parentctx = self._ctx
-        _parentState = self.state
-        localctx = LaTeXParser.ExpContext(self, self._ctx, _parentState)
-        _prevctx = localctx
-        _startState = 30
-        self.enterRecursionRule(localctx, 30, self.RULE_exp, _p)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 185
-            self.comp()
-            self._ctx.stop = self._input.LT(-1)
-            self.state = 201
-            self._errHandler.sync(self)
-            _alt = self._interp.adaptivePredict(self._input,16,self._ctx)
-            while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
-                if _alt==1:
-                    if self._parseListeners is not None:
-                        self.triggerExitRuleEvent()
-                    _prevctx = localctx
-                    localctx = LaTeXParser.ExpContext(self, _parentctx, _parentState)
-                    self.pushNewRecursionContext(localctx, _startState, self.RULE_exp)
-                    self.state = 187
-                    if not self.precpred(self._ctx, 2):
-                        from antlr4.error.Errors import FailedPredicateException
-                        raise FailedPredicateException(self, "self.precpred(self._ctx, 2)")
-                    self.state = 188
-                    self.match(LaTeXParser.CARET)
-                    self.state = 194
-                    self._errHandler.sync(self)
-                    token = self._input.LA(1)
-                    if token in [LaTeXParser.CMD_MATHIT, LaTeXParser.DIFFERENTIAL, LaTeXParser.LETTER, LaTeXParser.NUMBER, LaTeXParser.SYMBOL]:
-                        self.state = 189
-                        self.atom()
-                        pass
-                    elif token in [LaTeXParser.L_BRACE]:
-                        self.state = 190
-                        self.match(LaTeXParser.L_BRACE)
-                        self.state = 191
-                        self.expr()
-                        self.state = 192
-                        self.match(LaTeXParser.R_BRACE)
-                        pass
-                    else:
-                        raise NoViableAltException(self)
-
-                    self.state = 197
-                    self._errHandler.sync(self)
-                    la_ = self._interp.adaptivePredict(self._input,15,self._ctx)
-                    if la_ == 1:
-                        self.state = 196
-                        self.subexpr()
-
-             
-                self.state = 203
-                self._errHandler.sync(self)
-                _alt = self._interp.adaptivePredict(self._input,16,self._ctx)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.unrollRecursionContexts(_parentctx)
-        return localctx
-
-    class Exp_nofuncContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Exp_nofuncContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def comp_nofunc(self):
-            return self.getTypedRuleContext(LaTeXParser.Comp_nofuncContext,0)
-
-
-        def exp_nofunc(self):
-            return self.getTypedRuleContext(LaTeXParser.Exp_nofuncContext,0)
-
-
-        def CARET(self):
-            return self.getToken(LaTeXParser.CARET, 0)
-
-        def atom(self):
-            return self.getTypedRuleContext(LaTeXParser.AtomContext,0)
-
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def subexpr(self):
-            return self.getTypedRuleContext(LaTeXParser.SubexprContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_exp_nofunc
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterExp_nofunc"):
-                listener.enterExp_nofunc(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitExp_nofunc"):
-                listener.exitExp_nofunc(self)
-
-
-
-    def exp_nofunc(self, _p=0):
-        _parentctx = self._ctx
-        _parentState = self.state
-        localctx = LaTeXParser.Exp_nofuncContext(self, self._ctx, _parentState)
-        _prevctx = localctx
-        _startState = 32
-        self.enterRecursionRule(localctx, 32, self.RULE_exp_nofunc, _p)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 205
-            self.comp_nofunc()
-            self._ctx.stop = self._input.LT(-1)
-            self.state = 221
-            self._errHandler.sync(self)
-            _alt = self._interp.adaptivePredict(self._input,19,self._ctx)
-            while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
-                if _alt==1:
-                    if self._parseListeners is not None:
-                        self.triggerExitRuleEvent()
-                    _prevctx = localctx
-                    localctx = LaTeXParser.Exp_nofuncContext(self, _parentctx, _parentState)
-                    self.pushNewRecursionContext(localctx, _startState, self.RULE_exp_nofunc)
-                    self.state = 207
-                    if not self.precpred(self._ctx, 2):
-                        from antlr4.error.Errors import FailedPredicateException
-                        raise FailedPredicateException(self, "self.precpred(self._ctx, 2)")
-                    self.state = 208
-                    self.match(LaTeXParser.CARET)
-                    self.state = 214
-                    self._errHandler.sync(self)
-                    token = self._input.LA(1)
-                    if token in [LaTeXParser.CMD_MATHIT, LaTeXParser.DIFFERENTIAL, LaTeXParser.LETTER, LaTeXParser.NUMBER, LaTeXParser.SYMBOL]:
-                        self.state = 209
-                        self.atom()
-                        pass
-                    elif token in [LaTeXParser.L_BRACE]:
-                        self.state = 210
-                        self.match(LaTeXParser.L_BRACE)
-                        self.state = 211
-                        self.expr()
-                        self.state = 212
-                        self.match(LaTeXParser.R_BRACE)
-                        pass
-                    else:
-                        raise NoViableAltException(self)
-
-                    self.state = 217
-                    self._errHandler.sync(self)
-                    la_ = self._interp.adaptivePredict(self._input,18,self._ctx)
-                    if la_ == 1:
-                        self.state = 216
-                        self.subexpr()
-
-             
-                self.state = 223
-                self._errHandler.sync(self)
-                _alt = self._interp.adaptivePredict(self._input,19,self._ctx)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.unrollRecursionContexts(_parentctx)
-        return localctx
-
-    class CompContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.CompContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def group(self):
-            return self.getTypedRuleContext(LaTeXParser.GroupContext,0)
-
-
-        def abs_group(self):
-            return self.getTypedRuleContext(LaTeXParser.Abs_groupContext,0)
-
-
-        def func(self):
-            return self.getTypedRuleContext(LaTeXParser.FuncContext,0)
-
-
-        def atom(self):
-            return self.getTypedRuleContext(LaTeXParser.AtomContext,0)
-
-
-        def frac(self):
-            return self.getTypedRuleContext(LaTeXParser.FracContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_comp
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterComp"):
-                listener.enterComp(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitComp"):
-                listener.exitComp(self)
-
-
-
-
-    def comp(self):
-
-        localctx = LaTeXParser.CompContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 34, self.RULE_comp)
-        try:
-            self.state = 229
-            self._errHandler.sync(self)
-            la_ = self._interp.adaptivePredict(self._input,20,self._ctx)
-            if la_ == 1:
-                self.enterOuterAlt(localctx, 1)
-                self.state = 224
-                self.group()
-                pass
-
-            elif la_ == 2:
-                self.enterOuterAlt(localctx, 2)
-                self.state = 225
-                self.abs_group()
-                pass
-
-            elif la_ == 3:
-                self.enterOuterAlt(localctx, 3)
-                self.state = 226
-                self.func()
-                pass
-
-            elif la_ == 4:
-                self.enterOuterAlt(localctx, 4)
-                self.state = 227
-                self.atom()
-                pass
-
-            elif la_ == 5:
-                self.enterOuterAlt(localctx, 5)
-                self.state = 228
-                self.frac()
-                pass
-
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Comp_nofuncContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Comp_nofuncContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def group(self):
-            return self.getTypedRuleContext(LaTeXParser.GroupContext,0)
-
-
-        def abs_group(self):
-            return self.getTypedRuleContext(LaTeXParser.Abs_groupContext,0)
-
-
-        def atom(self):
-            return self.getTypedRuleContext(LaTeXParser.AtomContext,0)
-
-
-        def frac(self):
-            return self.getTypedRuleContext(LaTeXParser.FracContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_comp_nofunc
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterComp_nofunc"):
-                listener.enterComp_nofunc(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitComp_nofunc"):
-                listener.exitComp_nofunc(self)
-
-
-
-
-    def comp_nofunc(self):
-
-        localctx = LaTeXParser.Comp_nofuncContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 36, self.RULE_comp_nofunc)
-        try:
-            self.state = 235
-            self._errHandler.sync(self)
-            token = self._input.LA(1)
-            if token in [LaTeXParser.L_PAREN, LaTeXParser.L_BRACE, LaTeXParser.L_BRACKET]:
-                self.enterOuterAlt(localctx, 1)
-                self.state = 231
-                self.group()
-                pass
-            elif token in [LaTeXParser.BAR]:
-                self.enterOuterAlt(localctx, 2)
-                self.state = 232
-                self.abs_group()
-                pass
-            elif token in [LaTeXParser.CMD_MATHIT, LaTeXParser.DIFFERENTIAL, LaTeXParser.LETTER, LaTeXParser.NUMBER, LaTeXParser.SYMBOL]:
-                self.enterOuterAlt(localctx, 3)
-                self.state = 233
-                self.atom()
-                pass
-            elif token in [LaTeXParser.CMD_FRAC]:
-                self.enterOuterAlt(localctx, 4)
-                self.state = 234
-                self.frac()
-                pass
-            else:
-                raise NoViableAltException(self)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class GroupContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.GroupContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def L_PAREN(self):
-            return self.getToken(LaTeXParser.L_PAREN, 0)
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def R_PAREN(self):
-            return self.getToken(LaTeXParser.R_PAREN, 0)
-
-        def L_BRACKET(self):
-            return self.getToken(LaTeXParser.L_BRACKET, 0)
-
-        def R_BRACKET(self):
-            return self.getToken(LaTeXParser.R_BRACKET, 0)
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_group
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterGroup"):
-                listener.enterGroup(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitGroup"):
-                listener.exitGroup(self)
-
-
-
-
-    def group(self):
-
-        localctx = LaTeXParser.GroupContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 38, self.RULE_group)
-        try:
-            self.state = 249
-            self._errHandler.sync(self)
-            token = self._input.LA(1)
-            if token in [LaTeXParser.L_PAREN]:
-                self.enterOuterAlt(localctx, 1)
-                self.state = 237
-                self.match(LaTeXParser.L_PAREN)
-                self.state = 238
-                self.expr()
-                self.state = 239
-                self.match(LaTeXParser.R_PAREN)
-                pass
-            elif token in [LaTeXParser.L_BRACKET]:
-                self.enterOuterAlt(localctx, 2)
-                self.state = 241
-                self.match(LaTeXParser.L_BRACKET)
-                self.state = 242
-                self.expr()
-                self.state = 243
-                self.match(LaTeXParser.R_BRACKET)
-                pass
-            elif token in [LaTeXParser.L_BRACE]:
-                self.enterOuterAlt(localctx, 3)
-                self.state = 245
-                self.match(LaTeXParser.L_BRACE)
-                self.state = 246
-                self.expr()
-                self.state = 247
-                self.match(LaTeXParser.R_BRACE)
-                pass
-            else:
-                raise NoViableAltException(self)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Abs_groupContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Abs_groupContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def BAR(self, i=None):
-            if i is None:
-                return self.getTokens(LaTeXParser.BAR)
-            else:
-                return self.getToken(LaTeXParser.BAR, i)
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_abs_group
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterAbs_group"):
-                listener.enterAbs_group(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitAbs_group"):
-                listener.exitAbs_group(self)
-
-
-
-
-    def abs_group(self):
-
-        localctx = LaTeXParser.Abs_groupContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 40, self.RULE_abs_group)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 251
-            self.match(LaTeXParser.BAR)
-            self.state = 252
-            self.expr()
-            self.state = 253
-            self.match(LaTeXParser.BAR)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class AtomContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.AtomContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def LETTER(self):
-            return self.getToken(LaTeXParser.LETTER, 0)
-
-        def SYMBOL(self):
-            return self.getToken(LaTeXParser.SYMBOL, 0)
-
-        def subexpr(self):
-            return self.getTypedRuleContext(LaTeXParser.SubexprContext,0)
-
-
-        def NUMBER(self):
-            return self.getToken(LaTeXParser.NUMBER, 0)
-
-        def DIFFERENTIAL(self):
-            return self.getToken(LaTeXParser.DIFFERENTIAL, 0)
-
-        def mathit(self):
-            return self.getTypedRuleContext(LaTeXParser.MathitContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_atom
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterAtom"):
-                listener.enterAtom(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitAtom"):
-                listener.exitAtom(self)
-
-
-
-
-    def atom(self):
-
-        localctx = LaTeXParser.AtomContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 42, self.RULE_atom)
-        self._la = 0 # Token type
-        try:
-            self.state = 262
-            self._errHandler.sync(self)
-            token = self._input.LA(1)
-            if token in [LaTeXParser.LETTER, LaTeXParser.SYMBOL]:
-                self.enterOuterAlt(localctx, 1)
-                self.state = 255
-                _la = self._input.LA(1)
-                if not(_la==LaTeXParser.LETTER or _la==LaTeXParser.SYMBOL):
-                    self._errHandler.recoverInline(self)
-                else:
-                    self._errHandler.reportMatch(self)
-                    self.consume()
-                self.state = 257
-                self._errHandler.sync(self)
-                la_ = self._interp.adaptivePredict(self._input,23,self._ctx)
-                if la_ == 1:
-                    self.state = 256
-                    self.subexpr()
-
-
-                pass
-            elif token in [LaTeXParser.NUMBER]:
-                self.enterOuterAlt(localctx, 2)
-                self.state = 259
-                self.match(LaTeXParser.NUMBER)
-                pass
-            elif token in [LaTeXParser.DIFFERENTIAL]:
-                self.enterOuterAlt(localctx, 3)
-                self.state = 260
-                self.match(LaTeXParser.DIFFERENTIAL)
-                pass
-            elif token in [LaTeXParser.CMD_MATHIT]:
-                self.enterOuterAlt(localctx, 4)
-                self.state = 261
-                self.mathit()
-                pass
-            else:
-                raise NoViableAltException(self)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class MathitContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.MathitContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def CMD_MATHIT(self):
-            return self.getToken(LaTeXParser.CMD_MATHIT, 0)
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def mathit_text(self):
-            return self.getTypedRuleContext(LaTeXParser.Mathit_textContext,0)
-
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_mathit
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterMathit"):
-                listener.enterMathit(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitMathit"):
-                listener.exitMathit(self)
-
-
-
-
-    def mathit(self):
-
-        localctx = LaTeXParser.MathitContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 44, self.RULE_mathit)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 264
-            self.match(LaTeXParser.CMD_MATHIT)
-            self.state = 265
-            self.match(LaTeXParser.L_BRACE)
-            self.state = 266
-            self.mathit_text()
-            self.state = 267
-            self.match(LaTeXParser.R_BRACE)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Mathit_textContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Mathit_textContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def LETTER(self, i=None):
-            if i is None:
-                return self.getTokens(LaTeXParser.LETTER)
-            else:
-                return self.getToken(LaTeXParser.LETTER, i)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_mathit_text
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterMathit_text"):
-                listener.enterMathit_text(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitMathit_text"):
-                listener.exitMathit_text(self)
-
-
-
-
-    def mathit_text(self):
-
-        localctx = LaTeXParser.Mathit_textContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 46, self.RULE_mathit_text)
-        self._la = 0 # Token type
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 272
-            self._errHandler.sync(self)
-            _la = self._input.LA(1)
-            while _la==LaTeXParser.LETTER:
-                self.state = 269
-                self.match(LaTeXParser.LETTER)
-                self.state = 274
-                self._errHandler.sync(self)
-                _la = self._input.LA(1)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class FracContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.FracContext, self).__init__(parent, invokingState)
-            self.parser = parser
-            self.upper = None # ExprContext
-            self.lower = None # ExprContext
-
-        def CMD_FRAC(self):
-            return self.getToken(LaTeXParser.CMD_FRAC, 0)
-
-        def L_BRACE(self, i=None):
-            if i is None:
-                return self.getTokens(LaTeXParser.L_BRACE)
-            else:
-                return self.getToken(LaTeXParser.L_BRACE, i)
-
-        def R_BRACE(self, i=None):
-            if i is None:
-                return self.getTokens(LaTeXParser.R_BRACE)
-            else:
-                return self.getToken(LaTeXParser.R_BRACE, i)
-
-        def expr(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.ExprContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.ExprContext,i)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_frac
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterFrac"):
-                listener.enterFrac(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitFrac"):
-                listener.exitFrac(self)
-
-
-
-
-    def frac(self):
-
-        localctx = LaTeXParser.FracContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 48, self.RULE_frac)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 275
-            self.match(LaTeXParser.CMD_FRAC)
-            self.state = 276
-            self.match(LaTeXParser.L_BRACE)
-            self.state = 277
-            localctx.upper = self.expr()
-            self.state = 278
-            self.match(LaTeXParser.R_BRACE)
-            self.state = 279
-            self.match(LaTeXParser.L_BRACE)
-            self.state = 280
-            localctx.lower = self.expr()
-            self.state = 281
-            self.match(LaTeXParser.R_BRACE)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Func_normalContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Func_normalContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def FUNC_LOG(self):
-            return self.getToken(LaTeXParser.FUNC_LOG, 0)
-
-        def FUNC_LN(self):
-            return self.getToken(LaTeXParser.FUNC_LN, 0)
-
-        def FUNC_SIN(self):
-            return self.getToken(LaTeXParser.FUNC_SIN, 0)
-
-        def FUNC_COS(self):
-            return self.getToken(LaTeXParser.FUNC_COS, 0)
-
-        def FUNC_TAN(self):
-            return self.getToken(LaTeXParser.FUNC_TAN, 0)
-
-        def FUNC_CSC(self):
-            return self.getToken(LaTeXParser.FUNC_CSC, 0)
-
-        def FUNC_SEC(self):
-            return self.getToken(LaTeXParser.FUNC_SEC, 0)
-
-        def FUNC_COT(self):
-            return self.getToken(LaTeXParser.FUNC_COT, 0)
-
-        def FUNC_ARCSIN(self):
-            return self.getToken(LaTeXParser.FUNC_ARCSIN, 0)
-
-        def FUNC_ARCCOS(self):
-            return self.getToken(LaTeXParser.FUNC_ARCCOS, 0)
-
-        def FUNC_ARCTAN(self):
-            return self.getToken(LaTeXParser.FUNC_ARCTAN, 0)
-
-        def FUNC_ARCCSC(self):
-            return self.getToken(LaTeXParser.FUNC_ARCCSC, 0)
-
-        def FUNC_ARCSEC(self):
-            return self.getToken(LaTeXParser.FUNC_ARCSEC, 0)
-
-        def FUNC_ARCCOT(self):
-            return self.getToken(LaTeXParser.FUNC_ARCCOT, 0)
-
-        def FUNC_SINH(self):
-            return self.getToken(LaTeXParser.FUNC_SINH, 0)
-
-        def FUNC_COSH(self):
-            return self.getToken(LaTeXParser.FUNC_COSH, 0)
-
-        def FUNC_TANH(self):
-            return self.getToken(LaTeXParser.FUNC_TANH, 0)
-
-        def FUNC_ARSINH(self):
-            return self.getToken(LaTeXParser.FUNC_ARSINH, 0)
-
-        def FUNC_ARCOSH(self):
-            return self.getToken(LaTeXParser.FUNC_ARCOSH, 0)
-
-        def FUNC_ARTANH(self):
-            return self.getToken(LaTeXParser.FUNC_ARTANH, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_func_normal
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterFunc_normal"):
-                listener.enterFunc_normal(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitFunc_normal"):
-                listener.exitFunc_normal(self)
-
-
-
-
-    def func_normal(self):
-
-        localctx = LaTeXParser.Func_normalContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 50, self.RULE_func_normal)
-        self._la = 0 # Token type
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 283
-            _la = self._input.LA(1)
-            if not((((_la) & ~0x3f) == 0 and ((1 << _la) & ((1 << LaTeXParser.FUNC_LOG) | (1 << LaTeXParser.FUNC_LN) | (1 << LaTeXParser.FUNC_SIN) | (1 << LaTeXParser.FUNC_COS) | (1 << LaTeXParser.FUNC_TAN) | (1 << LaTeXParser.FUNC_CSC) | (1 << LaTeXParser.FUNC_SEC) | (1 << LaTeXParser.FUNC_COT) | (1 << LaTeXParser.FUNC_ARCSIN) | (1 << LaTeXParser.FUNC_ARCCOS) | (1 << LaTeXParser.FUNC_ARCTAN) | (1 << LaTeXParser.FUNC_ARCCSC) | (1 << LaTeXParser.FUNC_ARCSEC) | (1 << LaTeXParser.FUNC_ARCCOT) | (1 << LaTeXParser.FUNC_SINH) | (1 << LaTeXParser.FUNC_COSH) | (1 << LaTeXParser.FUNC_TANH) | (1 << LaTeXParser.FUNC_ARSINH) | (1 << LaTeXParser.FUNC_ARCOSH) | (1 << LaTeXParser.FUNC_ARTANH))) != 0)):
-                self._errHandler.recoverInline(self)
-            else:
-                self._errHandler.reportMatch(self)
-                self.consume()
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class FuncContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.FuncContext, self).__init__(parent, invokingState)
-            self.parser = parser
-            self.root = None # ExprContext
-            self.base = None # ExprContext
-
-        def func_normal(self):
-            return self.getTypedRuleContext(LaTeXParser.Func_normalContext,0)
-
-
-        def L_PAREN(self):
-            return self.getToken(LaTeXParser.L_PAREN, 0)
-
-        def func_arg(self):
-            return self.getTypedRuleContext(LaTeXParser.Func_argContext,0)
-
-
-        def R_PAREN(self):
-            return self.getToken(LaTeXParser.R_PAREN, 0)
-
-        def func_arg_noparens(self):
-            return self.getTypedRuleContext(LaTeXParser.Func_arg_noparensContext,0)
-
-
-        def subexpr(self):
-            return self.getTypedRuleContext(LaTeXParser.SubexprContext,0)
-
-
-        def supexpr(self):
-            return self.getTypedRuleContext(LaTeXParser.SupexprContext,0)
-
-
-        def args(self):
-            return self.getTypedRuleContext(LaTeXParser.ArgsContext,0)
-
-
-        def LETTER(self):
-            return self.getToken(LaTeXParser.LETTER, 0)
-
-        def SYMBOL(self):
-            return self.getToken(LaTeXParser.SYMBOL, 0)
-
-        def FUNC_INT(self):
-            return self.getToken(LaTeXParser.FUNC_INT, 0)
-
-        def DIFFERENTIAL(self):
-            return self.getToken(LaTeXParser.DIFFERENTIAL, 0)
-
-        def frac(self):
-            return self.getTypedRuleContext(LaTeXParser.FracContext,0)
-
-
-        def additive(self):
-            return self.getTypedRuleContext(LaTeXParser.AdditiveContext,0)
-
-
-        def FUNC_SQRT(self):
-            return self.getToken(LaTeXParser.FUNC_SQRT, 0)
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def expr(self, i=None):
-            if i is None:
-                return self.getTypedRuleContexts(LaTeXParser.ExprContext)
-            else:
-                return self.getTypedRuleContext(LaTeXParser.ExprContext,i)
-
-
-        def L_BRACKET(self):
-            return self.getToken(LaTeXParser.L_BRACKET, 0)
-
-        def R_BRACKET(self):
-            return self.getToken(LaTeXParser.R_BRACKET, 0)
-
-        def mp(self):
-            return self.getTypedRuleContext(LaTeXParser.MpContext,0)
-
-
-        def FUNC_SUM(self):
-            return self.getToken(LaTeXParser.FUNC_SUM, 0)
-
-        def FUNC_PROD(self):
-            return self.getToken(LaTeXParser.FUNC_PROD, 0)
-
-        def subeq(self):
-            return self.getTypedRuleContext(LaTeXParser.SubeqContext,0)
-
-
-        def FUNC_LIM(self):
-            return self.getToken(LaTeXParser.FUNC_LIM, 0)
-
-        def limit_sub(self):
-            return self.getTypedRuleContext(LaTeXParser.Limit_subContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_func
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterFunc"):
-                listener.enterFunc(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitFunc"):
-                listener.exitFunc(self)
-
-
-
-
-    def func(self):
-
-        localctx = LaTeXParser.FuncContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 52, self.RULE_func)
-        self._la = 0 # Token type
-        try:
-            self.state = 358
-            self._errHandler.sync(self)
-            token = self._input.LA(1)
-            if token in [LaTeXParser.FUNC_LOG, LaTeXParser.FUNC_LN, LaTeXParser.FUNC_SIN, LaTeXParser.FUNC_COS, LaTeXParser.FUNC_TAN, LaTeXParser.FUNC_CSC, LaTeXParser.FUNC_SEC, LaTeXParser.FUNC_COT, LaTeXParser.FUNC_ARCSIN, LaTeXParser.FUNC_ARCCOS, LaTeXParser.FUNC_ARCTAN, LaTeXParser.FUNC_ARCCSC, LaTeXParser.FUNC_ARCSEC, LaTeXParser.FUNC_ARCCOT, LaTeXParser.FUNC_SINH, LaTeXParser.FUNC_COSH, LaTeXParser.FUNC_TANH, LaTeXParser.FUNC_ARSINH, LaTeXParser.FUNC_ARCOSH, LaTeXParser.FUNC_ARTANH]:
-                self.enterOuterAlt(localctx, 1)
-                self.state = 285
-                self.func_normal()
-                self.state = 298
-                self._errHandler.sync(self)
-                la_ = self._interp.adaptivePredict(self._input,30,self._ctx)
-                if la_ == 1:
-                    self.state = 287
-                    self._errHandler.sync(self)
-                    _la = self._input.LA(1)
-                    if _la==LaTeXParser.UNDERSCORE:
-                        self.state = 286
-                        self.subexpr()
-
-
-                    self.state = 290
-                    self._errHandler.sync(self)
-                    _la = self._input.LA(1)
-                    if _la==LaTeXParser.CARET:
-                        self.state = 289
-                        self.supexpr()
-
-
-                    pass
-
-                elif la_ == 2:
-                    self.state = 293
-                    self._errHandler.sync(self)
-                    _la = self._input.LA(1)
-                    if _la==LaTeXParser.CARET:
-                        self.state = 292
-                        self.supexpr()
-
-
-                    self.state = 296
-                    self._errHandler.sync(self)
-                    _la = self._input.LA(1)
-                    if _la==LaTeXParser.UNDERSCORE:
-                        self.state = 295
-                        self.subexpr()
-
-
-                    pass
-
-
-                self.state = 305
-                self._errHandler.sync(self)
-                la_ = self._interp.adaptivePredict(self._input,31,self._ctx)
-                if la_ == 1:
-                    self.state = 300
-                    self.match(LaTeXParser.L_PAREN)
-                    self.state = 301
-                    self.func_arg()
-                    self.state = 302
-                    self.match(LaTeXParser.R_PAREN)
-                    pass
-
-                elif la_ == 2:
-                    self.state = 304
-                    self.func_arg_noparens()
-                    pass
-
-
-                pass
-            elif token in [LaTeXParser.LETTER, LaTeXParser.SYMBOL]:
-                self.enterOuterAlt(localctx, 2)
-                self.state = 307
-                _la = self._input.LA(1)
-                if not(_la==LaTeXParser.LETTER or _la==LaTeXParser.SYMBOL):
-                    self._errHandler.recoverInline(self)
-                else:
-                    self._errHandler.reportMatch(self)
-                    self.consume()
-                self.state = 309
-                self._errHandler.sync(self)
-                _la = self._input.LA(1)
-                if _la==LaTeXParser.UNDERSCORE:
-                    self.state = 308
-                    self.subexpr()
-
-
-                self.state = 311
-                self.match(LaTeXParser.L_PAREN)
-                self.state = 312
-                self.args()
-                self.state = 313
-                self.match(LaTeXParser.R_PAREN)
-                pass
-            elif token in [LaTeXParser.FUNC_INT]:
-                self.enterOuterAlt(localctx, 3)
-                self.state = 315
-                self.match(LaTeXParser.FUNC_INT)
-                self.state = 322
-                self._errHandler.sync(self)
-                token = self._input.LA(1)
-                if token in [LaTeXParser.UNDERSCORE]:
-                    self.state = 316
-                    self.subexpr()
-                    self.state = 317
-                    self.supexpr()
-                    pass
-                elif token in [LaTeXParser.CARET]:
-                    self.state = 319
-                    self.supexpr()
-                    self.state = 320
-                    self.subexpr()
-                    pass
-                elif token in [LaTeXParser.ADD, LaTeXParser.SUB, LaTeXParser.L_PAREN, LaTeXParser.L_BRACE, LaTeXParser.L_BRACKET, LaTeXParser.BAR, LaTeXParser.FUNC_LIM, LaTeXParser.FUNC_INT, LaTeXParser.FUNC_SUM, LaTeXParser.FUNC_PROD, LaTeXParser.FUNC_LOG, LaTeXParser.FUNC_LN, LaTeXParser.FUNC_SIN, LaTeXParser.FUNC_COS, LaTeXParser.FUNC_TAN, LaTeXParser.FUNC_CSC, LaTeXParser.FUNC_SEC, LaTeXParser.FUNC_COT, LaTeXParser.FUNC_ARCSIN, LaTeXParser.FUNC_ARCCOS, LaTeXParser.FUNC_ARCTAN, LaTeXParser.FUNC_ARCCSC, LaTeXParser.FUNC_ARCSEC, LaTeXParser.FUNC_ARCCOT, LaTeXParser.FUNC_SINH, LaTeXParser.FUNC_COSH, LaTeXParser.FUNC_TANH, LaTeXParser.FUNC_ARSINH, LaTeXParser.FUNC_ARCOSH, LaTeXParser.FUNC_ARTANH, LaTeXParser.FUNC_SQRT, LaTeXParser.CMD_FRAC, LaTeXParser.CMD_MATHIT, LaTeXParser.DIFFERENTIAL, LaTeXParser.LETTER, LaTeXParser.NUMBER, LaTeXParser.SYMBOL]:
-                    pass
-                else:
-                    pass
-                self.state = 330
-                self._errHandler.sync(self)
-                la_ = self._interp.adaptivePredict(self._input,35,self._ctx)
-                if la_ == 1:
-                    self.state = 325
-                    self._errHandler.sync(self)
-                    la_ = self._interp.adaptivePredict(self._input,34,self._ctx)
-                    if la_ == 1:
-                        self.state = 324
-                        self.additive(0)
-
-
-                    self.state = 327
-                    self.match(LaTeXParser.DIFFERENTIAL)
-                    pass
-
-                elif la_ == 2:
-                    self.state = 328
-                    self.frac()
-                    pass
-
-                elif la_ == 3:
-                    self.state = 329
-                    self.additive(0)
-                    pass
-
-
-                pass
-            elif token in [LaTeXParser.FUNC_SQRT]:
-                self.enterOuterAlt(localctx, 4)
-                self.state = 332
-                self.match(LaTeXParser.FUNC_SQRT)
-                self.state = 337
-                self._errHandler.sync(self)
-                _la = self._input.LA(1)
-                if _la==LaTeXParser.L_BRACKET:
-                    self.state = 333
-                    self.match(LaTeXParser.L_BRACKET)
-                    self.state = 334
-                    localctx.root = self.expr()
-                    self.state = 335
-                    self.match(LaTeXParser.R_BRACKET)
-
-
-                self.state = 339
-                self.match(LaTeXParser.L_BRACE)
-                self.state = 340
-                localctx.base = self.expr()
-                self.state = 341
-                self.match(LaTeXParser.R_BRACE)
-                pass
-            elif token in [LaTeXParser.FUNC_SUM, LaTeXParser.FUNC_PROD]:
-                self.enterOuterAlt(localctx, 5)
-                self.state = 343
-                _la = self._input.LA(1)
-                if not(_la==LaTeXParser.FUNC_SUM or _la==LaTeXParser.FUNC_PROD):
-                    self._errHandler.recoverInline(self)
-                else:
-                    self._errHandler.reportMatch(self)
-                    self.consume()
-                self.state = 350
-                self._errHandler.sync(self)
-                token = self._input.LA(1)
-                if token in [LaTeXParser.UNDERSCORE]:
-                    self.state = 344
-                    self.subeq()
-                    self.state = 345
-                    self.supexpr()
-                    pass
-                elif token in [LaTeXParser.CARET]:
-                    self.state = 347
-                    self.supexpr()
-                    self.state = 348
-                    self.subeq()
-                    pass
-                else:
-                    raise NoViableAltException(self)
-
-                self.state = 352
-                self.mp(0)
-                pass
-            elif token in [LaTeXParser.FUNC_LIM]:
-                self.enterOuterAlt(localctx, 6)
-                self.state = 354
-                self.match(LaTeXParser.FUNC_LIM)
-                self.state = 355
-                self.limit_sub()
-                self.state = 356
-                self.mp(0)
-                pass
-            else:
-                raise NoViableAltException(self)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class ArgsContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.ArgsContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def args(self):
-            return self.getTypedRuleContext(LaTeXParser.ArgsContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_args
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterArgs"):
-                listener.enterArgs(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitArgs"):
-                listener.exitArgs(self)
-
-
-
-
-    def args(self):
-
-        localctx = LaTeXParser.ArgsContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 54, self.RULE_args)
-        try:
-            self.state = 365
-            self._errHandler.sync(self)
-            la_ = self._interp.adaptivePredict(self._input,39,self._ctx)
-            if la_ == 1:
-                self.enterOuterAlt(localctx, 1)
-                self.state = 360
-                self.expr()
-                self.state = 361
-                self.match(LaTeXParser.T__0)
-                self.state = 362
-                self.args()
-                pass
-
-            elif la_ == 2:
-                self.enterOuterAlt(localctx, 2)
-                self.state = 364
-                self.expr()
-                pass
-
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Limit_subContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Limit_subContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def UNDERSCORE(self):
-            return self.getToken(LaTeXParser.UNDERSCORE, 0)
-
-        def L_BRACE(self, i=None):
-            if i is None:
-                return self.getTokens(LaTeXParser.L_BRACE)
-            else:
-                return self.getToken(LaTeXParser.L_BRACE, i)
-
-        def LIM_APPROACH_SYM(self):
-            return self.getToken(LaTeXParser.LIM_APPROACH_SYM, 0)
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def R_BRACE(self, i=None):
-            if i is None:
-                return self.getTokens(LaTeXParser.R_BRACE)
-            else:
-                return self.getToken(LaTeXParser.R_BRACE, i)
-
-        def LETTER(self):
-            return self.getToken(LaTeXParser.LETTER, 0)
-
-        def SYMBOL(self):
-            return self.getToken(LaTeXParser.SYMBOL, 0)
-
-        def CARET(self):
-            return self.getToken(LaTeXParser.CARET, 0)
-
-        def ADD(self):
-            return self.getToken(LaTeXParser.ADD, 0)
-
-        def SUB(self):
-            return self.getToken(LaTeXParser.SUB, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_limit_sub
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterLimit_sub"):
-                listener.enterLimit_sub(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitLimit_sub"):
-                listener.exitLimit_sub(self)
-
-
-
-
-    def limit_sub(self):
-
-        localctx = LaTeXParser.Limit_subContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 56, self.RULE_limit_sub)
-        self._la = 0 # Token type
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 367
-            self.match(LaTeXParser.UNDERSCORE)
-            self.state = 368
-            self.match(LaTeXParser.L_BRACE)
-            self.state = 369
-            _la = self._input.LA(1)
-            if not(_la==LaTeXParser.LETTER or _la==LaTeXParser.SYMBOL):
-                self._errHandler.recoverInline(self)
-            else:
-                self._errHandler.reportMatch(self)
-                self.consume()
-            self.state = 370
-            self.match(LaTeXParser.LIM_APPROACH_SYM)
-            self.state = 371
-            self.expr()
-            self.state = 376
-            self._errHandler.sync(self)
-            _la = self._input.LA(1)
-            if _la==LaTeXParser.CARET:
-                self.state = 372
-                self.match(LaTeXParser.CARET)
-                self.state = 373
-                self.match(LaTeXParser.L_BRACE)
-                self.state = 374
-                _la = self._input.LA(1)
-                if not(_la==LaTeXParser.ADD or _la==LaTeXParser.SUB):
-                    self._errHandler.recoverInline(self)
-                else:
-                    self._errHandler.reportMatch(self)
-                    self.consume()
-                self.state = 375
-                self.match(LaTeXParser.R_BRACE)
-
-
-            self.state = 378
-            self.match(LaTeXParser.R_BRACE)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Func_argContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Func_argContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def func_arg(self):
-            return self.getTypedRuleContext(LaTeXParser.Func_argContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_func_arg
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterFunc_arg"):
-                listener.enterFunc_arg(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitFunc_arg"):
-                listener.exitFunc_arg(self)
-
-
-
-
-    def func_arg(self):
-
-        localctx = LaTeXParser.Func_argContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 58, self.RULE_func_arg)
-        try:
-            self.state = 385
-            self._errHandler.sync(self)
-            la_ = self._interp.adaptivePredict(self._input,41,self._ctx)
-            if la_ == 1:
-                self.enterOuterAlt(localctx, 1)
-                self.state = 380
-                self.expr()
-                pass
-
-            elif la_ == 2:
-                self.enterOuterAlt(localctx, 2)
-                self.state = 381
-                self.expr()
-                self.state = 382
-                self.match(LaTeXParser.T__0)
-                self.state = 383
-                self.func_arg()
-                pass
-
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class Func_arg_noparensContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.Func_arg_noparensContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def mp_nofunc(self):
-            return self.getTypedRuleContext(LaTeXParser.Mp_nofuncContext,0)
-
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_func_arg_noparens
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterFunc_arg_noparens"):
-                listener.enterFunc_arg_noparens(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitFunc_arg_noparens"):
-                listener.exitFunc_arg_noparens(self)
-
-
-
-
-    def func_arg_noparens(self):
-
-        localctx = LaTeXParser.Func_arg_noparensContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 60, self.RULE_func_arg_noparens)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 387
-            self.mp_nofunc(0)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class SubexprContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.SubexprContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def UNDERSCORE(self):
-            return self.getToken(LaTeXParser.UNDERSCORE, 0)
-
-        def atom(self):
-            return self.getTypedRuleContext(LaTeXParser.AtomContext,0)
-
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_subexpr
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterSubexpr"):
-                listener.enterSubexpr(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitSubexpr"):
-                listener.exitSubexpr(self)
-
-
-
-
-    def subexpr(self):
-
-        localctx = LaTeXParser.SubexprContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 62, self.RULE_subexpr)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 389
-            self.match(LaTeXParser.UNDERSCORE)
-            self.state = 395
-            self._errHandler.sync(self)
-            token = self._input.LA(1)
-            if token in [LaTeXParser.CMD_MATHIT, LaTeXParser.DIFFERENTIAL, LaTeXParser.LETTER, LaTeXParser.NUMBER, LaTeXParser.SYMBOL]:
-                self.state = 390
-                self.atom()
-                pass
-            elif token in [LaTeXParser.L_BRACE]:
-                self.state = 391
-                self.match(LaTeXParser.L_BRACE)
-                self.state = 392
-                self.expr()
-                self.state = 393
-                self.match(LaTeXParser.R_BRACE)
-                pass
-            else:
-                raise NoViableAltException(self)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class SupexprContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.SupexprContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def CARET(self):
-            return self.getToken(LaTeXParser.CARET, 0)
-
-        def atom(self):
-            return self.getTypedRuleContext(LaTeXParser.AtomContext,0)
-
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def expr(self):
-            return self.getTypedRuleContext(LaTeXParser.ExprContext,0)
-
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_supexpr
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterSupexpr"):
-                listener.enterSupexpr(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitSupexpr"):
-                listener.exitSupexpr(self)
-
-
-
-
-    def supexpr(self):
-
-        localctx = LaTeXParser.SupexprContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 64, self.RULE_supexpr)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 397
-            self.match(LaTeXParser.CARET)
-            self.state = 403
-            self._errHandler.sync(self)
-            token = self._input.LA(1)
-            if token in [LaTeXParser.CMD_MATHIT, LaTeXParser.DIFFERENTIAL, LaTeXParser.LETTER, LaTeXParser.NUMBER, LaTeXParser.SYMBOL]:
-                self.state = 398
-                self.atom()
-                pass
-            elif token in [LaTeXParser.L_BRACE]:
-                self.state = 399
-                self.match(LaTeXParser.L_BRACE)
-                self.state = 400
-                self.expr()
-                self.state = 401
-                self.match(LaTeXParser.R_BRACE)
-                pass
-            else:
-                raise NoViableAltException(self)
-
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class SubeqContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.SubeqContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def UNDERSCORE(self):
-            return self.getToken(LaTeXParser.UNDERSCORE, 0)
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def equality(self):
-            return self.getTypedRuleContext(LaTeXParser.EqualityContext,0)
-
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_subeq
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterSubeq"):
-                listener.enterSubeq(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitSubeq"):
-                listener.exitSubeq(self)
-
-
-
-
-    def subeq(self):
-
-        localctx = LaTeXParser.SubeqContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 66, self.RULE_subeq)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 405
-            self.match(LaTeXParser.UNDERSCORE)
-            self.state = 406
-            self.match(LaTeXParser.L_BRACE)
-            self.state = 407
-            self.equality()
-            self.state = 408
-            self.match(LaTeXParser.R_BRACE)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-    class SupeqContext(ParserRuleContext):
-
-        def __init__(self, parser, parent=None, invokingState=-1):
-            super(LaTeXParser.SupeqContext, self).__init__(parent, invokingState)
-            self.parser = parser
-
-        def UNDERSCORE(self):
-            return self.getToken(LaTeXParser.UNDERSCORE, 0)
-
-        def L_BRACE(self):
-            return self.getToken(LaTeXParser.L_BRACE, 0)
-
-        def equality(self):
-            return self.getTypedRuleContext(LaTeXParser.EqualityContext,0)
-
-
-        def R_BRACE(self):
-            return self.getToken(LaTeXParser.R_BRACE, 0)
-
-        def getRuleIndex(self):
-            return LaTeXParser.RULE_supeq
-
-        def enterRule(self, listener):
-            if hasattr(listener, "enterSupeq"):
-                listener.enterSupeq(self)
-
-        def exitRule(self, listener):
-            if hasattr(listener, "exitSupeq"):
-                listener.exitSupeq(self)
-
-
-
-
-    def supeq(self):
-
-        localctx = LaTeXParser.SupeqContext(self, self._ctx, self.state)
-        self.enterRule(localctx, 68, self.RULE_supeq)
-        try:
-            self.enterOuterAlt(localctx, 1)
-            self.state = 410
-            self.match(LaTeXParser.UNDERSCORE)
-            self.state = 411
-            self.match(LaTeXParser.L_BRACE)
-            self.state = 412
-            self.equality()
-            self.state = 413
-            self.match(LaTeXParser.R_BRACE)
-        except RecognitionException as re:
-            localctx.exception = re
-            self._errHandler.reportError(self, re)
-            self._errHandler.recover(self, re)
-        finally:
-            self.exitRule()
-        return localctx
-
-
-
-    def sempred(self, localctx, ruleIndex, predIndex):
-        if self._predicates == None:
-            self._predicates = dict()
-        self._predicates[1] = self.relation_sempred
-        self._predicates[4] = self.additive_sempred
-        self._predicates[5] = self.mp_sempred
-        self._predicates[6] = self.mp_nofunc_sempred
-        self._predicates[15] = self.exp_sempred
-        self._predicates[16] = self.exp_nofunc_sempred
-        pred = self._predicates.get(ruleIndex, None)
-        if pred is None:
-            raise Exception("No predicate with index:" + str(ruleIndex))
-        else:
-            return pred(localctx, predIndex)
-
-    def relation_sempred(self, localctx, predIndex):
-            if predIndex == 0:
-                return self.precpred(self._ctx, 2)
-         
-
-    def additive_sempred(self, localctx, predIndex):
-            if predIndex == 1:
-                return self.precpred(self._ctx, 2)
-         
-
-    def mp_sempred(self, localctx, predIndex):
-            if predIndex == 2:
-                return self.precpred(self._ctx, 2)
-         
-
-    def mp_nofunc_sempred(self, localctx, predIndex):
-            if predIndex == 3:
-                return self.precpred(self._ctx, 2)
-         
-
-    def exp_sempred(self, localctx, predIndex):
-            if predIndex == 4:
-                return self.precpred(self._ctx, 2)
-         
-
-    def exp_nofunc_sempred(self, localctx, predIndex):
-            if predIndex == 5:
-                return self.precpred(self._ctx, 2)
-         
-
-
-
-
